Casos de COVID-19 en Mexico

casos Calcula el numero de casos registrados por fecha agrupando (o sin hacerlo) por diferentes covariables. Por default calcula el total de casos (con y sin prueba positiva)

Uso

casos(
  datos_covid,
  entidades = c("AGUASCALIENTES", "BAJA CALIFORNIA", "BAJA CALIFORNIA SUR", "CAMPECHE",
    "CHIAPAS", "CHIHUAHUA", "CIUDAD DE MÉXICO", "COAHUILA DE ZARAGOZA", "COLIMA",
    "DURANGO", "GUANAJUATO", "GUERRERO", "HIDALGO", "JALISCO", "MÉXICO",
    "MICHOACÁN DE OCAMPO", "MORELOS", "NAYARIT", "NUEVO LEÓN", "OAXACA", "PUEBLA",
    "QUERÉTARO", "QUINTANA ROO", "SAN LUIS POTOSÍ", "SINALOA", "SONORA", "TABASCO",
    "TAMAULIPAS", "TLAXCALA", "VERACRUZ DE IGNACIO DE LA LLAVE", "YUCATÁN", "ZACATECAS"),
  group_by_entidad = TRUE,
  entidad_tipo = c("Unidad Medica", "Residencia", "Nacimiento"),
  fecha_tipo = c("Sintomas", "Ingreso", "Defuncion"),
  tipo_clasificacion = c("Sospechosos", "Confirmados COVID", "Negativo a COVID",
    "Inválido", "No realizado"),
  group_by_tipo_clasificacion = FALSE,
  tipo_paciente = c("AMBULATORIO", "HOSPITALIZADO", "NO ESPECIFICADO"),
  group_by_tipo_paciente = FALSE,
  tipo_uci = c("SI", "NO", "NO APLICA", "SE IGNORA", "NO ESPECIFICADO"),
  group_by_tipo_uci = FALSE,
  tipo_sector = c("CRUZ ROJA", "DIF", "ESTATAL", "IMSS", "IMSS-BIENESTAR", "ISSSTE",
    "MUNICIPAL", "PEMEX", "PRIVADA", "SEDENA", "SEMAR", "SSA", "UNIVERSITARIO",
    "NO ESPECIFICADO"),
  group_by_tipo_sector = FALSE,
  defunciones = FALSE,
  edad_cut = NULL,
  as_tibble = TRUE,
  fill_zeros = as_tibble,
  list_name = "casos",
  .grouping_vars = c()
)

Argumentos

datos_covid

(obligatorio) Lista de tibbles o duckdbs resultante de descarga_datos_abiertos() o read_datos_abiertos()

entidades

(opcional) Vector con las entidades de las unidades medicas a analizar. Opciones: AGUASCALIENTES, BAJA CALIFORNIA, BAJA CALIFORNIA SUR, CAMPECHE, CHIAPAS, CHIHUAHUA, CIUDAD DE MEXICO, COAHUILA DE ZARAGOZA , COLIMA, DURANGO, GUANAJUATO, GUERRERO, HIDALGO, JALISCO, MEXICO, MICHOACAN DE OCAMPO, MORELOS,NAYARIT NUEVO LEON, OAXACA ,PUEBLA, QUERETARO,QUINTANA ROO, SAN LUIS POTOSI, SINALOA, SONORA, TABASCO, TAMAULIPAS,TLAXCALA, VERACRUZ DE IGNACIO DE LA LLAVE, YUCATAN, ZACATECAS.

group_by_entidad

(opcional) TRUE obtiene los casos para cada entidad reportando en cada fecha la entidad y los casos en dicha entidad. FALSE junta las entidades sumando sus casos en una sola observacion por cada fecha.

entidad_tipo

(opcional) Indica a que se refiere las entidades seleccionadas. Elige una de las opciones: Unidad Medica (entidad de la unidad medica), Nacimiento (entidad de origen del individuo) o Residencia (entidad donde reside el individuo).

fecha_tipo

(opcional) Selecciona si la fecha que se utiliza es la fecha de Ingreso (si aplica), la fecha de Sintomas o la de Defuncion (si aplica). El default es fecha de Sintomas.

tipo_clasificacion

(opcional) Vector con el tipo de clasificaciones (por la prueba) a incluir:Sospechosos,Confirmados COVID, Negativo a COVID, Inv\u00e1lido, No realizado

group_by_tipo_clasificacion

(opcional) Booleana determinando si regresa la base con cada entrada agrupada por tipo_clasificacion (es decir cada fecha se generan tantos observaciones como grupos de tipo de clasificación) en caso TRUE. Si FALSE suma todos los casos del tipo de clasificacion por fecha dando un solo numero por fecha. El defalt es FALSE.

tipo_paciente

(opcional) Vector con el tipo de pacientes a incluir. Opciones: AMBULATORIO, HOSPITALIZADO, NO ESPECIFICADO. Por default se incluyen todos.

group_by_tipo_paciente

(opcional) Booleana determinando (caso TRUE) si regresa la base con cada entrada agrupada por tipo_paciente (es decir cada fecha se genera un renglon para AMBULATORIO, un renglon para HOSPITALIZADO, etc) o bien si se suman todos los grupos y cada fecha reporta solo la suma de estos (estilo AMBULATORIO + HOSPITALIZADO segun las categorias de tipo_paciente) El default es FALSE.

tipo_uci

(opcional) Vector con el tipo de valores para Unidad de Cuidado Intensivo (UCI) a incluir: SI,NO,NO APLICA,SE IGNORA,NO ESPECIFICADO. Por default se incluyen todos.

group_by_tipo_uci

(opcional) Booleana. El caso TRUE determina si regresa la base con cada fecha teniendo diferentes renglones uno para cada tipo_uci (es decir cada fecha se generan tantos observaciones como grupos de tipo de UCI) o bien en una sola fecha se suman todos los tipos de UCI (FALSE). El default es FALSE.

tipo_sector

(opcional) Vector con los sectores del sistema de salud a incluir: CRUZ ROJA,DIF,ESTATAL,IMSS,IMSS-BIENESTAR,ISSSTE, MUNICIPAL,PEMEX, PRIVADA,SEDENA,SEMAR,SSA, UNIVERSITARIO,NO ESPECIFICADO. Por default se incluyen todos.

group_by_tipo_sector

(opcional) Booleana determina en el caso de TRUE si regresa la base con cada entrada agrupada por tipo_sector (es decir cada fecha tiene una entrada con los del IMSS, una entrada distinta con los de ISSSTE, etc) o bien en caso de FALSE se devuelve una sola entrada por fecha con la suma IMSS + ISSSTE + etc segun los sectores seleccionados. El default es FALSE.

defunciones

(opcional) Booleana si incluir sólo defunciones TRUE o a todos FALSE. El default es FALSE.

edad_cut

(opcional) Vector con secuencia de edades para hacer grupos. Por ejemplo edad_cut = c(0, 10, Inf) arma dos grupos de edad de 0 a 10 y de 10 a infinito o bien edad_cut = c(15, 20) deja sólo los registros entre 15 y 20 años. Por default es NULL y no arma grupos etarios.

as_tibble

(opcional) Regresar como tibble el resultado. En caso de que as_tibble sea FALSE se devuelve como conexion en duckdb. Se recomienda el default (tibble).

fill_zeros

(opcional) En caso de que el resultado sea un tibble regresa observaciones para todas las combinaciones de variables incluyendo como 0 aquellas fechas cuando no se observaron casos. En caso contrario no se incluyen las filas donde no se observaron casos.

list_name

(opcional) Asigna un nombre en la lista de datos a la base generada

.grouping_vars

(opcional) Vector de variables adicionales de agrupacion de los conteos. Por ejemplo si se agrega .grouping_vars = 'DIABETES' entonces para cada fecha habra dos conteos de casos uno de los que tienen diabetes y uno de los que no.

Valor

Une a la lista de datos_covid una nueva entrada de nombre list_name

(default: casos) con una base de datos (tibble o dbConnection) con los resultados agregados.

casos - Base de datos generara con los datos agregados (el nombre cambia si se usa list_name).

dict - Diccionario de datos

dats - Datos originales (conexion a duckdb o tibble)

disconnect - Función para desconectarte de duckdb

... - Cualquier otro elemento que ya existiera en datos_covid

Detalles

La función es un grupo de funciones de dplyr optimizadas para velocidad. Por ejemplo calcular los casos por entidad se hace lo siguiente

datos_covid |> casos()

es lo mismo que:

library(dplyr)
datos_covid$casos <- datos_covid$dats |>
    group_by(ENTIDAD_UM, FECHA_SINTOMAS) |>
    tally() |>
    left_join(datos_covid$dict$ENTIDAD_UM, by = c("ENTIDAD_UM" = "CLAVE_ENTIDAD"))

Elaboraciones mas complicadas en casos tienen su equivalente en dplyr por ejemplo:

datos_covid <- datos_covid |>
  casos(
    entidad_tipo = "Residencia",
    entidades = c("BAJA CALIFORNIA", "BAJA CALIFORNIA SUR"),
    group_by_tipo_clasificacion = FALSE,
    tipo_paciente = c("AMBULATORIO", "HOSPITALIZADO"),
    group_by_tipo_paciente = TRUE,
    list_name = "bajas"
  )

es equivalente a

datos_covid$bajas <- datos_covid$dats |>
   filter(ENTIDAD_RES == "02" | ENTIDAD_RES == "03") |> #BC/BCS
   filter(TIPO_PACIENTE == 1 | TIPO_PACIENTE == 2) |> #Ambulatorio/Hospitalizado
   group_by(FECHA_SINTOMAS, ENTIDAD_RES, TIPO_PACIENTE) |>
   tally() |>
   left_join(datos_covid$dict$ENTIDAD_RES, by = c("ENTIDAD_RES" = "CLAVE_ENTIDAD")) |>
   left_join(datos_covid$dict$PACIENTE, by = c("TIPO_PACIENTE" = "CLAVE"))

Ver también

Ejemplos

# Para el ejemplo usaremos los datos precargados pero tu puedes
# correr el ejemplo descargando informacion mas reciente:
# datos_covid <- descarga_datos_abiertos() #Sugerido

datos_covid <- datosabiertos # Datos precargados para el ejemplo

# Casos por entidad
datos_covid <- datos_covid |> casos()
head(datos_covid$casos)
#> # A tibble: 6 × 5
#>   FECHA_SINTOMAS      ENTIDAD_UM     n ENTIDAD_FEDERATIVA  ABREVIATURA
#>   <dttm>              <chr>      <int> <chr>               <chr>      
#> 1 2021-07-01 00:00:00 02           229 BAJA CALIFORNIA     BC         
#> 2 2021-07-01 00:00:00 03           794 BAJA CALIFORNIA SUR BS         
#> 3 2021-07-02 00:00:00 02           240 BAJA CALIFORNIA     BC         
#> 4 2021-07-02 00:00:00 03           535 BAJA CALIFORNIA SUR BS         
#> 5 2021-07-03 00:00:00 02           269 BAJA CALIFORNIA     BC         
#> 6 2021-07-03 00:00:00 03           646 BAJA CALIFORNIA SUR BS         

# Defunciones por entidad
datos_covid <- datos_covid |> casos(defunciones = TRUE, list_name = "defunciones")
head(datos_covid$defunciones)
#> # A tibble: 6 × 5
#>   FECHA_SINTOMAS      ENTIDAD_UM     n ENTIDAD_FEDERATIVA  ABREVIATURA
#>   <dttm>              <chr>      <int> <chr>               <chr>      
#> 1 2021-07-01 00:00:00 02             5 BAJA CALIFORNIA     BC         
#> 2 2021-07-01 00:00:00 03            37 BAJA CALIFORNIA SUR BS         
#> 3 2021-07-02 00:00:00 02             7 BAJA CALIFORNIA     BC         
#> 4 2021-07-02 00:00:00 03            11 BAJA CALIFORNIA SUR BS         
#> 5 2021-07-03 00:00:00 02             7 BAJA CALIFORNIA     BC         
#> 6 2021-07-03 00:00:00 03            22 BAJA CALIFORNIA SUR BS         

# Hospitalizados por entidad
datos_covid <- datos_covid |>
  casos(tipo_paciente = "HOSPITALIZADO", list_name = "hospitalizados")
head(datos_covid$hospitalizados)
#> # A tibble: 6 × 5
#>   FECHA_SINTOMAS      ENTIDAD_UM     n ENTIDAD_FEDERATIVA  ABREVIATURA
#>   <dttm>              <chr>      <int> <chr>               <chr>      
#> 1 2021-07-01 00:00:00 02            17 BAJA CALIFORNIA     BC         
#> 2 2021-07-01 00:00:00 03            85 BAJA CALIFORNIA SUR BS         
#> 3 2021-07-02 00:00:00 02            26 BAJA CALIFORNIA     BC         
#> 4 2021-07-02 00:00:00 03            35 BAJA CALIFORNIA SUR BS         
#> 5 2021-07-03 00:00:00 02            19 BAJA CALIFORNIA     BC         
#> 6 2021-07-03 00:00:00 03            44 BAJA CALIFORNIA SUR BS         

# UCI por entidad
datos_covid <- datos_covid |> casos(tipo_uci = "SI", list_name = "uci")
head(datos_covid$uci)
#> # A tibble: 6 × 5
#>   FECHA_SINTOMAS      ENTIDAD_UM     n ENTIDAD_FEDERATIVA  ABREVIATURA
#>   <dttm>              <chr>      <int> <chr>               <chr>      
#> 1 2021-07-01 00:00:00 03             6 BAJA CALIFORNIA SUR BS         
#> 2 2021-07-02 00:00:00 03             3 BAJA CALIFORNIA SUR BS         
#> 3 2021-07-03 00:00:00 02             2 BAJA CALIFORNIA     BC         
#> 4 2021-07-03 00:00:00 03             8 BAJA CALIFORNIA SUR BS         
#> 5 2021-07-04 00:00:00 02             1 BAJA CALIFORNIA     BC         
#> 6 2021-07-04 00:00:00 03             3 BAJA CALIFORNIA SUR BS         

# Solo pacientes IMSS
datos_covid <- datos_covid |> casos(tipo_sector = "IMSS", list_name = "imss")
head(datos_covid$imss)
#> # A tibble: 6 × 5
#>   FECHA_SINTOMAS      ENTIDAD_UM     n ENTIDAD_FEDERATIVA  ABREVIATURA
#>   <dttm>              <chr>      <int> <chr>               <chr>      
#> 1 2021-07-01 00:00:00 02           194 BAJA CALIFORNIA     BC         
#> 2 2021-07-01 00:00:00 03           554 BAJA CALIFORNIA SUR BS         
#> 3 2021-07-02 00:00:00 02           196 BAJA CALIFORNIA     BC         
#> 4 2021-07-02 00:00:00 03           378 BAJA CALIFORNIA SUR BS         
#> 5 2021-07-03 00:00:00 02           231 BAJA CALIFORNIA     BC         
#> 6 2021-07-03 00:00:00 03           408 BAJA CALIFORNIA SUR BS         

# Pacientes IMSS y PEMEX separados
datos_covid <- datos_covid |> casos(tipo_sector = c("IMSS", "PEMEX"), list_name = "imss_y_pemex")
head(datos_covid$imss_y_pemex)
#> # A tibble: 6 × 5
#>   FECHA_SINTOMAS      ENTIDAD_UM     n ENTIDAD_FEDERATIVA  ABREVIATURA
#>   <dttm>              <chr>      <int> <chr>               <chr>      
#> 1 2021-07-01 00:00:00 02           194 BAJA CALIFORNIA     BC         
#> 2 2021-07-01 00:00:00 03           554 BAJA CALIFORNIA SUR BS         
#> 3 2021-07-02 00:00:00 02           196 BAJA CALIFORNIA     BC         
#> 4 2021-07-02 00:00:00 03           378 BAJA CALIFORNIA SUR BS         
#> 5 2021-07-03 00:00:00 02           231 BAJA CALIFORNIA     BC         
#> 6 2021-07-03 00:00:00 03           408 BAJA CALIFORNIA SUR BS         

# Pacientes IMSS y PEMEX sumados
datos_covid <- datos_covid |>
  casos(
    tipo_sector = c("IMSS", "PEMEX"), list_name = "imss_+_pemex",
    group_by_tipo_sector = TRUE
  )
head(datos_covid$`imss_+_pemex`)
#> # A tibble: 6 × 7
#>   FECHA_SINTOMAS      ENTIDAD_UM SECTOR     n ENTIDAD_FEDERATIVA ABREV…¹ DESCR…²
#>   <dttm>              <chr>       <dbl> <int> <chr>              <chr>   <chr>  
#> 1 2021-07-01 00:00:00 02              4   194 BAJA CALIFORNIA    BC      IMSS   
#> 2 2021-07-01 00:00:00 03              4   554 BAJA CALIFORNIA S… BS      IMSS   
#> 3 2021-07-02 00:00:00 02              4   196 BAJA CALIFORNIA    BC      IMSS   
#> 4 2021-07-02 00:00:00 03              4   378 BAJA CALIFORNIA S… BS      IMSS   
#> 5 2021-07-03 00:00:00 02              4   231 BAJA CALIFORNIA    BC      IMSS   
#> 6 2021-07-03 00:00:00 03              4   408 BAJA CALIFORNIA S… BS      IMSS   
#> # … with abbreviated variable names ¹​ABREVIATURA, ²​DESCRIPCION_TIPO_SECTOR

# Solo los de BAJA CALIFORNIA
datos_covid <- datos_covid |>
  casos(entidades = c("BAJA CALIFORNIA"), list_name = "BC")
head(datos_covid$BC)
#> # A tibble: 6 × 5
#>   FECHA_SINTOMAS      ENTIDAD_UM     n ENTIDAD_FEDERATIVA ABREVIATURA
#>   <dttm>              <chr>      <int> <chr>              <chr>      
#> 1 2021-07-01 00:00:00 02           229 BAJA CALIFORNIA    BC         
#> 2 2021-07-02 00:00:00 02           240 BAJA CALIFORNIA    BC         
#> 3 2021-07-03 00:00:00 02           269 BAJA CALIFORNIA    BC         
#> 4 2021-07-04 00:00:00 02           308 BAJA CALIFORNIA    BC         
#> 5 2021-07-05 00:00:00 02           333 BAJA CALIFORNIA    BC         
#> 6 2021-07-06 00:00:00 02           317 BAJA CALIFORNIA    BC         

# Solo los de BAJA CALIFORNIA por residencia
datos_covid <- datos_covid |>
  casos(entidades = c("BAJA CALIFORNIA"), entidad_tipo = "Residencia", list_name = "residencia")
head(datos_covid$residencia)
#> # A tibble: 6 × 5
#>   FECHA_SINTOMAS      ENTIDAD_RES     n ENTIDAD_FEDERATIVA ABREVIATURA
#>   <dttm>              <chr>       <int> <chr>              <chr>      
#> 1 2021-07-01 00:00:00 02            230 BAJA CALIFORNIA    BC         
#> 2 2021-07-02 00:00:00 02            241 BAJA CALIFORNIA    BC         
#> 3 2021-07-03 00:00:00 02            266 BAJA CALIFORNIA    BC         
#> 4 2021-07-04 00:00:00 02            308 BAJA CALIFORNIA    BC         
#> 5 2021-07-05 00:00:00 02            331 BAJA CALIFORNIA    BC         
#> 6 2021-07-06 00:00:00 02            317 BAJA CALIFORNIA    BC         

# Agrupando casos por tipo de clasificacion
datos_covid <- datos_covid |>
  casos(
    entidades = c("BAJA CALIFORNIA", "BAJA CALIFORNIA SUR"),
    group_by_tipo_clasificacion = TRUE,
    list_name = "BC_BCS"
  )
head(datos_covid$BC_BCS)
#> # A tibble: 6 × 7
#>   FECHA_SINTOMAS      ENTIDAD_UM CLASIFICACION_F…¹     n ENTID…² ABREV…³ CLASI…⁴
#>   <dttm>              <chr>                  <dbl> <int> <chr>   <chr>   <chr>  
#> 1 2021-07-01 00:00:00 02                         1     1 BAJA C… BC      CASO D…
#> 2 2021-07-01 00:00:00 02                         3    39 BAJA C… BC      CASO D…
#> 3 2021-07-01 00:00:00 02                         6     8 BAJA C… BC      CASO S…
#> 4 2021-07-01 00:00:00 02                         7   181 BAJA C… BC      NEGATI…
#> 5 2021-07-01 00:00:00 03                         1     1 BAJA C… BS      CASO D…
#> 6 2021-07-01 00:00:00 03                         3   368 BAJA C… BS      CASO D…
#> # … with abbreviated variable names ¹​CLASIFICACION_FINAL, ²​ENTIDAD_FEDERATIVA,
#> #   ³​ABREVIATURA, ⁴​CLASIFICACIÓN

# Regresa la suma de los de BC + BCS por tipo de paciente
datos_covid <- datos_covid |>
  casos(
    entidades = c("BAJA CALIFORNIA", "BAJA CALIFORNIA SUR"),
    group_by_tipo_clasificacion = FALSE,
    tipo_paciente = c("AMBULATORIO", "HOSPITALIZADO"),
    group_by_tipo_paciente = TRUE,
    list_name = "BC_+_BCS"
  )
head(datos_covid$`BC_+_BCS`)
#> # A tibble: 6 × 7
#>   FECHA_SINTOMAS      ENTIDAD_UM TIPO_PACIENTE     n ENTIDAD_F…¹ ABREV…² DESCR…³
#>   <dttm>              <chr>              <dbl> <int> <chr>       <chr>   <chr>  
#> 1 2021-07-01 00:00:00 02                     1   212 BAJA CALIF… BC      AMBULA…
#> 2 2021-07-01 00:00:00 02                     2    17 BAJA CALIF… BC      HOSPIT…
#> 3 2021-07-01 00:00:00 03                     1   709 BAJA CALIF… BS      AMBULA…
#> 4 2021-07-01 00:00:00 03                     2    85 BAJA CALIF… BS      HOSPIT…
#> 5 2021-07-02 00:00:00 02                     1   214 BAJA CALIF… BC      AMBULA…
#> 6 2021-07-02 00:00:00 02                     2    26 BAJA CALIF… BC      HOSPIT…
#> # … with abbreviated variable names ¹​ENTIDAD_FEDERATIVA, ²​ABREVIATURA,
#> #   ³​DESCRIPCION_TIPO_PACIENTE

# Si deseas agrupar por una variable que no este en las opciones
datos_covid <- datos_covid |>
  casos(
    group_by_entidad = FALSE,
    tipo_paciente = c("AMBULATORIO", "HOSPITALIZADO"),
    group_by_tipo_paciente = TRUE,
    list_name = "sexo",
    .grouping_vars = c("SEXO")
  )
head(datos_covid$sexo)
#> # A tibble: 6 × 5
#>   FECHA_SINTOMAS       SEXO TIPO_PACIENTE     n DESCRIPCION_TIPO_PACIENTE
#>   <dttm>              <int>         <dbl> <int> <chr>                    
#> 1 2021-07-01 00:00:00     1             1   471 AMBULATORIO              
#> 2 2021-07-01 00:00:00     1             2    44 HOSPITALIZADO            
#> 3 2021-07-01 00:00:00     2             1   450 AMBULATORIO              
#> 4 2021-07-01 00:00:00     2             2    58 HOSPITALIZADO            
#> 5 2021-07-02 00:00:00     1             1   378 AMBULATORIO              
#> 6 2021-07-02 00:00:00     1             2    27 HOSPITALIZADO            

# Si no recuerdas la codificacion de los sexos puedes usar el diccionario:
datos_covid$sexo <- datos_covid$sexo |>
  dplyr::left_join(datos_covid$dict$SEXO, by = c("SEXO" = "CLAVE"))
head(datos_covid$sexo)
#> # A tibble: 6 × 6
#>   FECHA_SINTOMAS       SEXO TIPO_PACIENTE     n DESCRIPCION_TIPO_PACIE…¹ DESCR…²
#>   <dttm>              <dbl>         <dbl> <int> <chr>                    <chr>  
#> 1 2021-07-01 00:00:00     1             1   471 AMBULATORIO              MUJER  
#> 2 2021-07-01 00:00:00     1             2    44 HOSPITALIZADO            MUJER  
#> 3 2021-07-01 00:00:00     2             1   450 AMBULATORIO              HOMBRE 
#> 4 2021-07-01 00:00:00     2             2    58 HOSPITALIZADO            HOMBRE 
#> 5 2021-07-02 00:00:00     1             1   378 AMBULATORIO              MUJER  
#> 6 2021-07-02 00:00:00     1             2    27 HOSPITALIZADO            MUJER  
#> # … with abbreviated variable names ¹​DESCRIPCION_TIPO_PACIENTE, ²​DESCRIPCIÓN

# Si no recuerdas todas las variables de la base puedes usar glimpse para ver por
# que otras variables puedes clasificar
datos_covid$dats |> dplyr::glimpse()
#> Rows: 58,965
#> Columns: 40
#> $ FECHA_ACTUALIZACION   <dttm> 2022-08-09, 2022-08-09, 2022-08-09, 2022-08-09,…
#> $ ID_REGISTRO           <chr> "g164c1a", "78ce0b", "526821", "4ab051", "ba3171…
#> $ ORIGEN                <int> 2, 2, 1, 1, 1, 1, 2, 1, 2, 1, 1, 1, 1, 2, 1, 2, …
#> $ SECTOR                <int> 12, 12, 4, 4, 4, 6, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4…
#> $ ENTIDAD_UM            <chr> "02", "03", "03", "03", "03", "03", "03", "03", …
#> $ SEXO                  <int> 1, 1, 2, 2, 2, 2, 2, 2, 1, 1, 1, 1, 1, 1, 1, 1, …
#> $ ENTIDAD_NAC           <chr> "02", "09", "03", "25", "03", "03", "25", "03", …
#> $ ENTIDAD_RES           <chr> "02", "03", "03", "03", "03", "03", "03", "03", …
#> $ MUNICIPIO_RES         <chr> "001", "003", "008", "008", "008", "003", "003",…
#> $ TIPO_PACIENTE         <int> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 2, 1, 1, …
#> $ FECHA_INGRESO         <dttm> 2021-07-22, 2021-07-08, 2021-07-01, 2021-07-01,…
#> $ FECHA_SINTOMAS        <dttm> 2021-07-21, 2021-07-04, 2021-07-01, 2021-07-01,…
#> $ FECHA_DEF             <dttm> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,…
#> $ INTUBADO              <int> 97, 97, 97, 97, 97, 97, 97, 97, 97, 97, 97, 97, …
#> $ NEUMONIA              <int> 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 1, 2, 2, …
#> $ EDAD                  <int> 43, 17, 28, 30, 22, 32, 35, 24, 40, 21, 78, 13, …
#> $ NACIONALIDAD          <int> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, …
#> $ EMBARAZO              <int> 2, 2, 97, 97, 97, 97, 97, 97, 2, 2, 2, 2, 2, 2, …
#> $ HABLA_LENGUA_INDIG    <int> 2, 99, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2,…
#> $ INDIGENA              <int> 2, 99, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2,…
#> $ DIABETES              <int> 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 1, 2, 2, …
#> $ EPOC                  <int> 2, 2, 2, 2, 1, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, …
#> $ ASMA                  <int> 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, …
#> $ INMUSUPR              <int> 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, …
#> $ HIPERTENSION          <int> 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, …
#> $ OTRA_COM              <int> 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, …
#> $ CARDIOVASCULAR        <int> 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, …
#> $ OBESIDAD              <int> 2, 2, 1, 1, 2, 2, 1, 2, 2, 2, 2, 2, 2, 2, 2, 2, …
#> $ RENAL_CRONICA         <int> 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, …
#> $ TABAQUISMO            <int> 2, 2, 2, 2, 1, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, …
#> $ OTRO_CASO             <int> 2, 2, 1, 1, 1, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, …
#> $ TOMA_MUESTRA_LAB      <int> 2, 1, 2, 2, 2, 2, 1, 2, 2, 2, 2, 2, 2, 1, 2, 2, …
#> $ RESULTADO_LAB         <int> 97, 2, 97, 97, 97, 97, 1, 97, 97, 97, 97, 97, 97…
#> $ TOMA_MUESTRA_ANTIGENO <int> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, …
#> $ RESULTADO_ANTIGENO    <int> 1, 2, 2, 1, 2, 2, 1, 1, 1, 2, 2, 1, 2, 2, 2, 2, …
#> $ CLASIFICACION_FINAL   <int> 3, 7, 7, 3, 7, 7, 3, 3, 3, 7, 7, 3, 7, 6, 7, 7, …
#> $ MIGRANTE              <int> 99, 99, 99, 99, 99, 99, 99, 99, 99, 99, 99, 99, …
#> $ PAIS_NACIONALIDAD     <chr> "México", "México", "México", "México", "México"…
#> $ PAIS_ORIGEN           <chr> "97", "97", "97", "97", "97", "97", "97", "97", …
#> $ UCI                   <int> 97, 97, 97, 97, 97, 97, 97, 97, 97, 97, 97, 97, …

#Una vez hayas concluido tu trabajo no olvides desconectar
datos_covid$disconnect()
#> ✔ Desconectado