DESCARGA BASE DE DATOS ABIERTOS DE LA DIRECCION GENERAL DE EPIDEMIOLOGIA

descarga_datos_abiertos downloads and labels all COVID-19 data for Mexico from the Direccion General de EpidemiologIa at https://www.gob.mx/salud/documentos/datos-abiertos-152127.

descarga_datos_abiertos descarga y etiqueta todos los datos de COVID-19 de Mexico de la Direccion General de EpidemiologIa en https://www.gob.mx/salud/documentos/datos-abiertos-152127.

Usage

descarga_datos_abiertos(
  download_method = "curl",
  file_download_data = tempfile(),
  file_download_dictionary = tempfile(),
  remove_zip_after_download = TRUE,
  quiet = FALSE,
  parse_dictionary = TRUE,
  parse_warnings = FALSE,
  language = "English"
)

Arguments

download_method

Methods for download file (default = "curl"). Other options are "internal", "wininet" (Windows) "libcurl", "wget", "curl". See download.file / Metodos para descargar el archivo (default = "curl"). Otras opciones incluyen "interno","wininet" (Windows) " libcurl", "wget", "curl". Vease download.file.

file_download_data

Name of file to save the data / Nombre del archivo para guardar los datos.

file_download_dictionary

Name of file to save the dictionary / Nombre del archivo donde guardar el diccioanrio.

remove_zip_after_download

If the downloaded zip file should be saved / Si los archivos zip descargados deben ser almacenados.

quiet

Show messages? / Mostrar mensajes?

parse_dictionary

Change labels to values of dictionary? Default = TRUE else the data is downloaded and presented as is. / Se agregan etiquetas a partir del diccionario? Default = TRUE en caso contrario solo se presentan los datos descargados como estan.

parse_warnings

If show parsing errors when reading csv / Si muestra errores cuando lee el csv

language

('Espanol', 'English') Message languages/ Idiomas del mensaje.

Details

While reading the data it is perfectly normal to get a Warning for parsing: the dataset has issues. En la lectura de datos es normal tener un Warning por parsing: la base viene chueca.

This is not an official product / este no es un producto oficial

Examples

if (FALSE) {
datos_covid <- descarga_datos_abiertos(language = "Espanol")
}