LEE LA BASE DE DATOS DE OCUPACION HOSPITALARIA DE RED IRAG

descarga_datos_red_irag Lee los datos de ocupacion hospitalaria de la RED IRAG disponibles en https://www.gits.igg.unam.mx/red-irag-dashboard/reviewHome#

Uso

descarga_datos_red_irag(
  nivel = c("Estatal", "Unidad Médica"),
  cache = NULL,
  use_cache_on_failure = TRUE,
  quiet = TRUE,
  force_download = FALSE,
  show_warnings = TRUE,
  ...
)

Argumentos

nivel

"Estatal" o "Unidad Medica"

cache

cache para pins::board_url.

use_cache_on_failure

parametro para pins::board_url.

quiet

booleana para no imprimir mensajes en la consola.

force_download

analiza si cambio el pin y descarga datos nuevos en caso afirmativo

show_warnings

si arrojar warnings

...

parametros adicionales para pins::pin_download.

Valor

data.frame con los datos de ocupacion hospitalaria

Detalles

Los datos de Red IRAG son descargados diariamente de manera automatica en Github: RodrigoZepeda/CapacidadHospitalariaMX y esta funcion los lee de ahi

Ver también

Ejemplos

if (FALSE) {
# Descarga de datos estatales
ocupacion_hospitalaria <- descarga_datos_red_irag("Estatal")

# Descarga de datos por unidad medica
ocupacion_hospitalaria <- descarga_datos_red_irag("Unidad Medica")

# Si ya descargaste hace menos de un día el programa solito se da cuenta y lee de memoria
# sin verificar que el contenido en Internet haya cambiado
ocupacion_hospitalaria <- descarga_datos_red_irag("Unidad Medica")

# Puedes forzarlo a checar el contenido en Internet usando
variantes_covid <- descarga_datos_red_irag("Unidad Medica", force_download = TRUE)
}