Lee la base de datos de variantes de COVID-19 en Mexico generada por GISAID

descarga_datos_variantes_GISAID Lee los datos de variantes del reporte nacional diario en RodrigoZepeda/VariantesCovid creado a partir de la informacion de la Global Initiative on Sharing Avian Influenza Data (GISAID)

Uso

descarga_datos_variantes_GISAID(
  nivel = c("nacional", "cdmx"),
  cache = NULL,
  use_cache_on_failure = TRUE,
  quiet = FALSE,
  force_download = FALSE,
  show_warnings = TRUE,
  ...
)

Argumentos

nivel

(opcional) si se desea descargar informacion "nacional" (default) o de la Ciudad de Mexico: "cdmx".

cache

(opcional) cache para pins::board_url(). Representa el directorio donde se almacenaran los datos descargados en formato de pins.

use_cache_on_failure

(opcional) parametro para pins::board_url(). En caso de TRUE (default) si no puede descargar nueva informacion utiliza la que ya tiene en memoria aunque sea vieja.

quiet

(opcional) booleana para no imprimir mensajes en la consola.

force_download

(opcional) analiza si cambio el pin y descarga datos nuevos en caso afirmativo.

show_warnings

(opcional) si arrojar warnings o callar

...

parametros adicionales para pins::pin_download().

Valor

tibble con los datos de porcentuales de variantes

variant - La variante clasificada mediante Pangolin

semana - Semana epidemiologica lubridate::epiweek() a la que corresponde la variante

ano - Anio al que corresponde la toma de muestra

n - El total de muestras de dicha semana registradas para esa variante

freq - La proporcion de las variantes de dicha semana ocupada por dicha variante. Se obtiene dividiendo n/sum(n) para cada semana.

Actualizacion - La fecha de actualizacion ultima de los datos.

Fuente - La fuente desde la cual se obtuvo la informacion de dicha variante.

Detalles

Cada vez que uses estos datos necesitas citar a GISAID (ver referencias) asi como el reporte en RodrigoZepeda/VariantesCovid

Los datos son descargados de manera automatica en mi Github: RodrigoZepeda/VariantesCovid el programa descarga_datos_variantes_GISAID se conecta a dicho repositorio, busca si la informacion esta actualizada y si si la descarga, si no, utiliza informacion almacenada en el cache local.

La descarga usa el paquete pins

Referencias

Khare, S., et al (2021) GISAID's Role in Pandemic Response. China CDC Weekly, 3(49): 1049-1051. doi:10.46234/ccdcw2021.255 PMCID: 8668406

Elbe, S. and Buckland-Merrett, G. (2017) Data, disease and diplomacy: GISAID’s innovative contribution to global health. Global Challenges, 1:33-46. doi:10.1002/gch2.1018 PMCID: 31565258

Shu, Y. and McCauley, J. (2017) GISAID: from vision to reality. EuroSurveillance, 22(13) doi:10.2807/1560-7917.ES.2017.22.13.30494 PMCID: PMC5388101

Zepeda-Tello, R. (2022). Reporte Nacional de Variantes de COVID-19. URL: https://github.com/RodrigoZepeda/VariantesCovid

Ver también

Ejemplos

# Descarga de variantes a nivel nacional
url_global <- "https://raw.githubusercontent.com/RodrigoZepeda/VariantesCovid/main/tablas/"
if (RCurl::url.exists(paste0(url_global, "Proporcion_variantes_nacional.csv"))) {
  variantes_covid <- descarga_datos_variantes_GISAID("nacional")
}
#> → Descargando nacional desde <https://raw.githubusercontent.com/RodrigoZepeda/VariantesCovid/main/tablas/Proporcion_variantes_nacional.csv>
#> Warning: La descarga mas reciente fue hace 0.70552 dias. Como tiene menos de un dia
#> usare esa. Escribe `force_download = TRUE` si quieres descargar de todas
#> formas. Para desactivar este mensaje `show_warnings = FALSE.`

# Descarga de variantes para CDMX
if (RCurl::url.exists(paste0(url_global, "Proporcion_variantes_cdmx.csv"))) {
  variantes_covid <- descarga_datos_variantes_GISAID("cdmx")
}
#> → Descargando cdmx desde <https://raw.githubusercontent.com/RodrigoZepeda/VariantesCovid/main/tablas/Proporcion_variantes_cdmx.csv>
#> Warning: La descarga mas reciente fue hace 0.70552 dias. Como tiene menos de un dia
#> usare esa. Escribe `force_download = TRUE` si quieres descargar de todas
#> formas. Para desactivar este mensaje `show_warnings = FALSE.`