Numero de Pruebas

numero_pruebas Calcula el numero total de pruebas por fecha agrupando (o sin hacerlo) por covariables. Por default calcula la el numero de pruebas de antigeno y PCR por separado para cada estado.

Uso

numero_pruebas(
  datos_covid,
  entidades = c("AGUASCALIENTES", "BAJA CALIFORNIA", "BAJA CALIFORNIA SUR", "CAMPECHE",
    "CHIAPAS", "CHIHUAHUA", "CIUDAD DE MÉXICO", "COAHUILA DE ZARAGOZA", "COLIMA",
    "DURANGO", "GUANAJUATO", "GUERRERO", "HIDALGO", "JALISCO", "MÉXICO",
    "MICHOACÁN DE OCAMPO", "MORELOS", "NAYARIT", "NUEVO LEÓN", "OAXACA", "PUEBLA",
    "QUERÉTARO", "QUINTANA ROO", "SAN LUIS POTOSÍ", "SINALOA", "SONORA", "TABASCO",
    "TAMAULIPAS", "TLAXCALA", "VERACRUZ DE IGNACIO DE LA LLAVE", "YUCATÁN", "ZACATECAS"),
  group_by_entidad = TRUE,
  entidad_tipo = c("Unidad Medica", "Residencia", "Nacimiento"),
  fecha_tipo = c("Sintomas", "Ingreso", "Defuncion"),
  tipo_prueba = c("Antigeno", "PCR"),
  group_by_tipo_prueba = TRUE,
  tipo_paciente = c("AMBULATORIO", "HOSPITALIZADO", "NO ESPECIFICADO"),
  group_by_tipo_paciente = FALSE,
  tipo_uci = c("SI", "NO", "NO APLICA", "SE IGNORA", "NO ESPECIFICADO"),
  group_by_tipo_uci = FALSE,
  tipo_sector = c("CRUZ ROJA", "DIF", "ESTATAL", "IMSS", "IMSS-BIENESTAR", "ISSSTE",
    "MUNICIPAL", "PEMEX", "PRIVADA", "SEDENA", "SEMAR", "SSA", "UNIVERSITARIO",
    "NO ESPECIFICADO"),
  group_by_tipo_sector = FALSE,
  defunciones = FALSE,
  edad_cut = NULL,
  as_tibble = TRUE,
  fill_zeros = as_tibble,
  list_name = "numero_pruebas",
  .grouping_vars = c()
)

Argumentos

datos_covid

(obligatorio) Lista de tibbles o duckdbs resultante de descarga_datos_abiertos() o read_datos_abiertos()

entidades

(opcional) Vector con las entidades de las unidades medicas a analizar. Opciones: AGUASCALIENTES, BAJA CALIFORNIA, BAJA CALIFORNIA SUR, CAMPECHE, CHIAPAS, CHIHUAHUA, CIUDAD DE MEXICO, COAHUILA DE ZARAGOZA , COLIMA, DURANGO, GUANAJUATO, GUERRERO, HIDALGO, JALISCO, MEXICO, MICHOACAN DE OCAMPO, MORELOS,NAYARIT NUEVO LEON, OAXACA ,PUEBLA, QUERETARO,QUINTANA ROO, SAN LUIS POTOSI, SINALOA, SONORA, TABASCO, TAMAULIPAS,TLAXCALA, VERACRUZ DE IGNACIO DE LA LLAVE, YUCATAN, ZACATECAS.

group_by_entidad

(opcional) TRUE obtiene los casos para cada entidad reportando en cada fecha la entidad y los casos en dicha entidad. FALSE junta las entidades sumando sus casos en una sola observacion por cada fecha.

entidad_tipo

(opcional) Indica a que se refiere las entidades seleccionadas. Elige una de las opciones: Unidad Medica (entidad de la unidad medica), Nacimiento (entidad de origen del individuo) o Residencia (entidad donde reside el individuo).

fecha_tipo

(opcional) Selecciona si la fecha que se utiliza es la fecha de Ingreso (si aplica), la fecha de Sintomas o la de Defuncion (si aplica). El default es fecha de Sintomas.

tipo_prueba

(opcional) Vector con el tipo de pruebas a incluir Antigeno, PCR. Por default se incluyen ambas.

group_by_tipo_prueba

(opcional) Booleana determinando si regresa la base con cada entrada agrupada por tipo_prueba. En caso TRUE (cada fecha y entidad reporta separado el los casos de PCR y Antigeno). En caso FALSE se juntan los casos de PCR y Antigeno para devolver un unico numero por fecha.

tipo_paciente

(opcional) Vector con el tipo de pacientes a incluir. Opciones: AMBULATORIO, HOSPITALIZADO, NO ESPECIFICADO. Por default se incluyen todos.

group_by_tipo_paciente

(opcional) Booleana determinando (caso TRUE) si regresa la base con cada entrada agrupada por tipo_paciente (es decir cada fecha se genera un renglon para AMBULATORIO, un renglon para HOSPITALIZADO, etc) o bien si se suman todos los grupos y cada fecha reporta solo la suma de estos (estilo AMBULATORIO + HOSPITALIZADO segun las categorias de tipo_paciente) El default es FALSE.

tipo_uci

(opcional) Vector con el tipo de valores para Unidad de Cuidado Intensivo (UCI) a incluir: SI,NO,NO APLICA,SE IGNORA,NO ESPECIFICADO. Por default se incluyen todos.

group_by_tipo_uci

(opcional) Booleana. El caso TRUE determina si regresa la base con cada fecha teniendo diferentes renglones uno para cada tipo_uci (es decir cada fecha se generan tantos observaciones como grupos de tipo de UCI) o bien en una sola fecha se suman todos los tipos de UCI (FALSE). El default es FALSE.

tipo_sector

(opcional) Vector con los sectores del sistema de salud a incluir: CRUZ ROJA,DIF,ESTATAL,IMSS,IMSS-BIENESTAR,ISSSTE, MUNICIPAL,PEMEX, PRIVADA,SEDENA,SEMAR,SSA, UNIVERSITARIO,NO ESPECIFICADO. Por default se incluyen todos.

group_by_tipo_sector

(opcional) Booleana determina en el caso de TRUE si regresa la base con cada entrada agrupada por tipo_sector (es decir cada fecha tiene una entrada con los del IMSS, una entrada distinta con los de ISSSTE, etc) o bien en caso de FALSE se devuelve una sola entrada por fecha con la suma IMSS + ISSSTE + etc segun los sectores seleccionados. El default es FALSE.

defunciones

(opcional) Booleana si incluir sólo defunciones TRUE o a todos FALSE. El default es FALSE.

edad_cut

(opcional) Vector con secuencia de edades para hacer grupos. Por ejemplo edad_cut = c(0, 10, Inf) arma dos grupos de edad de 0 a 10 y de 10 a infinito o bien edad_cut = c(15, 20) deja sólo los registros entre 15 y 20 años. Por default es NULL y no arma grupos etarios.

as_tibble

(opcional) Regresar como tibble el resultado. En caso de que as_tibble sea FALSE se devuelve como conexion en duckdb. Se recomienda el default (tibble).

fill_zeros

(opcional) En caso de que el resultado sea un tibble regresa observaciones para todas las combinaciones de variables incluyendo como 0 aquellas fechas cuando no se observaron casos. En caso contrario no se incluyen las filas donde no se observaron casos.

list_name

(opcional) Asigna un nombre en la lista de datos a la base generada

.grouping_vars

(opcional) Vector de variables adicionales de agrupacion de los conteos. Por ejemplo si se agrega .grouping_vars = 'DIABETES' entonces para cada fecha habra dos conteos de casos uno de los que tienen diabetes y uno de los que no.

Valor

Adiciona a la lista de datos_covid una nueva entrada de nombre list_name

(default: numero_pruebas) con una base de datos (tibble o duckdb) con los resultados agregados.

numero_pruebas - Base de datos generara con los datos agregados (el nombre cambia si se usa list_name).

dict - Diccionario de datos

dats - Datos originales (conexion a duckdb o tibble)

disconnect - Función para desconectarte de duckdb

... - Cualquier otro elemento que ya existiera en datos_covid

Detalles

Las pruebas de PCR (polymerase chain reaction) identifican material genetico de un organismo (por ejemplo un virus como el COVID-19 o la influenza). Las pruebas de antigeno (o pruebas rapidas) detectan algunas proteinas que conforman el virus.

Para mas informacion sobre las pruebas y su interpretacion puedes consultar las guias del CDC

Ejemplos

# Para el ejemplo usaremos los datos precargados pero tu puedes
# correr el ejemplo descargando informacion mas reciente:
# datos_covid <- descarga_datos_abiertos() #Sugerido

datos_covid <- datosabiertos

# Número de pruebas PCR/Antigeno a nivel nacional por estado
datos_covid <- datos_covid |> numero_pruebas()
head(datos_covid$numero_pruebas)
#> # A tibble: 6 × 6
#>   FECHA_SINTOMAS      ENTIDAD_UM TIPO_PRUEBA     n ENTIDAD_FEDERATIVA  ABREVIA…¹
#>   <dttm>              <chr>      <chr>       <int> <chr>               <chr>    
#> 1 2021-07-01 00:00:00 02         ANTIGENO      125 BAJA CALIFORNIA     BC       
#> 2 2021-07-01 00:00:00 02         PCR            21 BAJA CALIFORNIA     BC       
#> 3 2021-07-01 00:00:00 03         ANTIGENO      479 BAJA CALIFORNIA SUR BS       
#> 4 2021-07-01 00:00:00 03         PCR            69 BAJA CALIFORNIA SUR BS       
#> 5 2021-07-02 00:00:00 02         ANTIGENO      143 BAJA CALIFORNIA     BC       
#> 6 2021-07-02 00:00:00 02         PCR            36 BAJA CALIFORNIA     BC       
#> # … with abbreviated variable name ¹​ABREVIATURA

# Número de pruebas nacionales pero sin separar por tipo ni estado
datos_covid <- datos_covid |>
  numero_pruebas(
    group_by_entidad = FALSE, group_by_tipo_prueba = FALSE,
    list_name = "Todas_las_pruebas"
  )
head(datos_covid$Todas_las_pruebas)
#> # A tibble: 6 × 2
#>   FECHA_SINTOMAS          n
#>   <dttm>              <int>
#> 1 2021-07-01 00:00:00   694
#> 2 2021-07-02 00:00:00   564
#> 3 2021-07-03 00:00:00   629
#> 4 2021-07-04 00:00:00   703
#> 5 2021-07-05 00:00:00   782
#> 6 2021-07-06 00:00:00   766

# Positivos en Baja California Sur
datos_covid <- datos_covid |>
  numero_pruebas(
    entidades = c("BAJA CALIFORNIA SUR"),
    list_name = "BCS"
  )
head(datos_covid$BCS)
#> # A tibble: 6 × 6
#>   FECHA_SINTOMAS      ENTIDAD_UM TIPO_PRUEBA     n ENTIDAD_FEDERATIVA  ABREVIA…¹
#>   <dttm>              <chr>      <chr>       <int> <chr>               <chr>    
#> 1 2021-07-01 00:00:00 03         ANTIGENO      479 BAJA CALIFORNIA SUR BS       
#> 2 2021-07-01 00:00:00 03         PCR            69 BAJA CALIFORNIA SUR BS       
#> 3 2021-07-02 00:00:00 03         ANTIGENO      331 BAJA CALIFORNIA SUR BS       
#> 4 2021-07-02 00:00:00 03         PCR            54 BAJA CALIFORNIA SUR BS       
#> 5 2021-07-03 00:00:00 03         ANTIGENO      387 BAJA CALIFORNIA SUR BS       
#> 6 2021-07-03 00:00:00 03         PCR            58 BAJA CALIFORNIA SUR BS       
#> # … with abbreviated variable name ¹​ABREVIATURA

# Si deseas agrupar por una variable que no este en las opciones asi como tipo paciente
datos_covid <- datos_covid |>
  numero_pruebas(
    tipo_paciente = c("AMBULATORIO", "HOSPITALIZADO"),
    group_by_tipo_paciente = TRUE,
    .grouping_vars = c("DIABETES"),
    list_name = "pruebas_diabetes"
  )
head(datos_covid$pruebas_diabetes)
#> # A tibble: 6 × 9
#>   FECHA_SINTOMAS      DIABETES ENTIDAD_UM TIPO_P…¹ TIPO_…²     n ENTID…³ ABREV…⁴
#>   <dttm>                 <int> <chr>      <chr>      <dbl> <int> <chr>   <chr>  
#> 1 2021-07-01 00:00:00        1 02         ANTIGENO       1     8 BAJA C… BC     
#> 2 2021-07-01 00:00:00        1 02         ANTIGENO       2     5 BAJA C… BC     
#> 3 2021-07-01 00:00:00        1 02         PCR            1     2 BAJA C… BC     
#> 4 2021-07-01 00:00:00        1 02         PCR            2     3 BAJA C… BC     
#> 5 2021-07-01 00:00:00        1 03         ANTIGENO       1    21 BAJA C… BS     
#> 6 2021-07-01 00:00:00        1 03         ANTIGENO       2     7 BAJA C… BS     
#> # … with 1 more variable: DESCRIPCION_TIPO_PACIENTE <chr>, and abbreviated
#> #   variable names ¹​TIPO_PRUEBA, ²​TIPO_PACIENTE, ³​ENTIDAD_FEDERATIVA,
#> #   ⁴​ABREVIATURA

# Una vez hayas concluido tu trabajo no olvides desconectar
datos_covid$disconnect()
#> ✔ Desconectado