PLOT DATOS DE COVID

plot_covid Grafica automaticamente la base de datos de covid

Uso

plot_covid(
  datos_covid,
  df_name = "casos",
  df_date_index = stringr::str_subset(colnames(datos_covid[df_name][[1]]),
    "FECHA|fecha|Fecha"),
  df_variable = NULL,
  df_covariates = NULL,
  facet_scale = "free_y",
  facet_ncol = 4,
  date_break_format = "2 months",
  date_labels_format = "%B-%y",
  type = c("point", "line", "spline", "area"),
  plot_theme = ggplot2::theme(panel.background = ggplot2::element_rect(fill = "white"),
    plot.background = ggplot2::element_rect(fill = "white"), axis.text.x =
    ggplot2::element_text(angle = 90, hjust = 1), axis.line.x =
    ggplot2::element_line(color = "black"), legend.position = "none"),
  ...
)

Argumentos

datos_covid

If no data is available it automatically downloads COVID-19 information y calcula los casos por entidad

df_name

Nombre de la base de datos dentro de la lista datos_covid

df_date_index

Nombre de la variable que contiene la fecha

df_variable

Nombre de la variable que se va a graficar en el eje y

df_covariates

Covariables para el facet (maximo 2)

facet_scale

Escala para el facet_wrap

facet_ncol

Numero de columnas para el facet_wrap

date_break_format

Breaks para el eje x

date_labels_format

Formato de fecha para el eje x

type

Tipo de grafica (line, area, spline o point)

plot_theme

Tema para el ggplot2

...

Parametros adicionales para ggformula::geom_spline en caso de elegir type="spline"

Detalles

Por default calcula el total de casos. This is not an official product / este no es un producto oficial

Ejemplos

if (FALSE) {
#Grafica de casos por entidad
datos_covid %>% casos() %>% plot_covid()

#Grafica de casos nacional
datos_covid %>% casos(group_by_entidad = F) %>% plot_covid()

#Grafica de casos nacional
datos_covid %>% casos(group_by_entidad = F) %>% plot_covid()

#Ajuste mediante splines
datos_covid %>% casos(group_by_entidad = F) %>% plot_covid(type = "spline", spar = 0.5)

#Ajuste mediante splines
datos_covid %>% casos(group_by_entidad = T, group_by_tipo_clasificacion = T, df_variable = "n") %>%
       plot_covid()
}