LEE LA BASE DE DATOS ABIERTOS DE LA DIRECCION GENERAL DE EPIDEMIOLOGIA QUE YA DESCARGASTE

read_datos_abiertos Lee los datos abiertos almacenados en tu base de MariaDB que bajaste con descarga_datos_abiertos. Intenta de manera automática determinar si los lee de MariaDB, csv ó zip

La funcion principal es read_datos_abiertos() la cual decide si los lee de zip, MARIADB o csv Tambien puedes usar las auxiliares respectivas

read_datos_abiertos_zip() Si sólo descargaste los datos de la DGE en .zip

read_datos_abiertos_csv() Si descargaste los datos de la DGE en .zip y los descomprimiste.

read_datos_abiertos_MariaDB() Si ya creaste tu table en MariaDB

Uso

read_datos_abiertos(
  datos_abiertos_path = NULL,
  driver = RMariaDB::MariaDB(),
  sqlimport = "mysqlimport",
  read_format = c("MariaDB", "tibble"),
  user = Sys.getenv("MariaDB_user"),
  password = Sys.getenv("MariaDB_password"),
  dbname = Sys.getenv("MariaDB_dbname"),
  host = Sys.getenv("MariaDB_host"),
  group = Sys.getenv("MariaDB_group"),
  port = Sys.getenv("MariaDB_port"),
  tblname = "covidmx",
  nthreads = max(parallel::detectCores() - 1, 1),
  ...
)

read_datos_abiertos_zip(
  datos_abiertos_zip_path,
  diccionario_zip_path = NULL,
  diccionario_unzipped_path = NULL,
  diccionario = NULL,
  read_format = c("MariaDB", "tibble"),
  driver = RMariaDB::MariaDB(),
  sqlimport = "mysqlimport",
  user = Sys.getenv("MariaDB_user"),
  password = Sys.getenv("MariaDB_password"),
  dbname = Sys.getenv("MariaDB_dbname"),
  host = Sys.getenv("MariaDB_host"),
  group = Sys.getenv("MariaDB_group"),
  port = Sys.getenv("MariaDB_port"),
  tblname = "covidmx",
  nthreads = max(parallel::detectCores() - 1, 1),
  download_process = c("pins", "download.file"),
  site.covid.dic = paste0("http://datosabiertos.salud.", "gob.mx/gobmx/salud/datos_a",
    "biertos/diccionario_datos_", "covid19.zip"),
  unzip_command = Sys.getenv("unzip_command"),
  unzip_args = Sys.getenv("unzip_args"),
  unzip_args_dict = list(exdir = ".", overwrite = TRUE),
  check_unzip_install = TRUE,
  clear_zip = download_process[1] != "pins",
  clear_csv = TRUE,
  use_dict = TRUE,
  quiet = FALSE,
  cache_datos = NULL,
  use_cache_on_failure = TRUE,
  cache_diccionario = NULL,
  force_download = FALSE,
  show_warnings = TRUE,
  board_url_name = "datos_abiertos",
  board_url_name_dict = "diccionario_covid",
  download_file_args = list(method = "curl", destfile = tempfile(), quiet = quiet),
  descarga_db_diccionario_ssa_args = list(),
  prestatement = paste0("SET sql_mode = 'NO_ENGINE_SUBSTITUTION,",
    "NO_AUTO_CREATE_USER';"),
  additional_mysqlimport_flags = c("--local"),
  load_data_infile_options = list(LOW_PRIORITY = "", CONCURRENT = "", LOCAL = "LOCAL",
    PARTITION = ""),
  ...
)

read_datos_abiertos_csv(
  datos_abiertos_unzipped_path,
  diccionario_zip_path = NULL,
  diccionario_unzipped_path = NULL,
  diccionario = NULL,
  read_format = c("MariaDB", "tibble"),
  driver = RMariaDB::MariaDB(),
  sqlimport = "mysqlimport",
  user = Sys.getenv("MariaDB_user"),
  password = Sys.getenv("MariaDB_password"),
  dbname = Sys.getenv("MariaDB_dbname"),
  host = Sys.getenv("MariaDB_host"),
  group = Sys.getenv("MariaDB_group"),
  port = Sys.getenv("MariaDB_port"),
  tblname = "covidmx",
  nthreads = max(parallel::detectCores() - 1, 1),
  download_process = c("pins", "download.file"),
  site.covid.dic = paste0("http://datosabiertos.salud.", "gob.mx/gobmx/salud/datos_a",
    "biertos/diccionario_datos_", "covid19.zip"),
  unzip_args_dict = list(exdir = ".", overwrite = TRUE),
  clear_csv = TRUE,
  quiet = FALSE,
  use_cache_on_failure = TRUE,
  cache_diccionario = NULL,
  force_download = FALSE,
  show_warnings = TRUE,
  board_url_name_dict = "diccionario_covid",
  download_file_args = list(method = "curl", destfile = tempfile(), quiet = quiet),
  descarga_db_diccionario_ssa_args = list(),
  prestatement = paste0("SET sql_mode = 'NO_ENGINE_SUBSTITUTION,",
    "NO_AUTO_CREATE_USER';"),
  additional_mysqlimport_flags = c("--local"),
  load_data_infile_options = list(LOW_PRIORITY = "", CONCURRENT = "", LOCAL = "LOCAL",
    PARTITION = ""),
  ...
)

read_datos_abiertos_MariaDB(
  user = Sys.getenv("MariaDB_user"),
  password = Sys.getenv("MariaDB_password"),
  dbname = Sys.getenv("MariaDB_dbname"),
  host = Sys.getenv("MariaDB_host"),
  group = Sys.getenv("MariaDB_group"),
  port = Sys.getenv("MariaDB_port"),
  driver = RMariaDB::MariaDB(),
  tblname = "covidmx",
  diccionario_zip_path = NULL,
  diccionario_unzipped_path = NULL,
  diccionario = NULL,
  download_process = c("pins", "download.file"),
  site.covid.dic = paste0("http://datosabiertos.salud.", "gob.mx/gobmx/salud/datos_a",
    "biertos/diccionario_datos_", "covid19.zip"),
  unzip_args_dict = list(exdir = ".", overwrite = TRUE),
  clear_zip = download_process[1] != "pins",
  clear_csv = TRUE,
  use_dict = TRUE,
  quiet = FALSE,
  use_cache_on_failure = TRUE,
  cache_diccionario = NULL,
  force_download = FALSE,
  show_warnings = TRUE,
  board_url_name_dict = "diccionario_covid",
  download_file_args = list(method = "curl", destfile = tempfile(), quiet = quiet),
  descarga_db_diccionario_ssa_args = list(),
  ...
)

Argumentos

datos_abiertos_path

Camino a los datos abiertos si son un zip o un csv o bien dejar en NULL si se quiere leer de MariaDB

driver

An SQL driver for dbConnect

sqlimport

An import command similar to mysqlimport

read_format

"MariaDB" o "tibble" establece el formato de lectura de la base de datos. En la mayoria de los casos "tibble" va a resultar en un error de memoria.

user

Usuario para dbConnect i.e. tu usuario de MariaDB. Puedes ponerlo como variable ambiental con Sys.setenv("MariaDB_user" = "tu_usuario")

password

password para dbConnect i.e. tu password de MariaDB Puedes ponerlo como variable ambiental con Sys.setenv("MariaDB_password" = "tu_passwodrd")

dbname

Nombre de la base de datos DATABASE paradbConnect puedes ponerlo con Sys.setenv("MariaDB_dbname" = "covidmx")

host

(opcional) Host para dbConnect. Si no sabes cual tienes seguro es localhost. Puedes ponerlo con Sys.setenv("MariaDB_host" = "localhost")

group

(opcional) Grupo de dbConnect i.e. tu grupo de MariaDB. Ponlo como variable ambiental usando Sys.setenv("MariaDB_group" = "tu_grupo")

port

Puerto de la conexion con MariaDB. Puedes ponerlo con Sys.setenv("MariaDB_group" = "tu_grupo")

tblname

Nombre de la tabla dentro de el DATABASE dbname donde guardar los datos por default se llama covidmx.

nthreads

Número de threads para escribir el archivo en MariaDB.

...

Parametros adicionales para pins::pin_download

datos_abiertos_zip_path

Camino a los datos abiertos si ya los descargaste en zip

diccionario_zip_path

Camino al diccionario si ya losdescargaste en zip

diccionario_unzipped_path

Camino al diccionario csv si ya lo descargaste y descomprimiste el archivo zip en un csv

diccionario

Lo que resulta de realizar una descarga del diccionario usando descarga_diccionario

download_process

(opcional) Metodo para descargar ya sea pins o download.file. Se recomienda pins pues guarda en memoria la fecha de la ultima descarga y analiza si ha pasado mas de un dia desde la descarga. En caso afirmativo verifica si el archivo ha cambiado y si hubo cambios entonces lo descarga.

site.covid.dic

Sitio para descarga del diccionario de datos.

unzip_command

Forma de extraer la base de datos de datos abiertos. La forma de llamarla es con system2(unzip_command, args = c(unzip_args, file_download_data)).

unzip_args

Argumentos de extraccion de la base de datos de datos abiertos. La forma de llamarla es con system2(unzip_command, args = c(unzip_args, file_download_data)).

unzip_args_dict

Lista de argumentos para usar utils::unzip en el diccionario de datos.

check_unzip_install

Bandera de verificacion para checar si tienes lo necesario para unzippear los datos.

clear_zip

Si borrar los archivos .zip descargados para el diccionario y los datos abiertos. No se recomienda si estas usando pins. Ve la nota para mas informacion.

clear_csv

Si borrar los archivos .csv que se generan despues de abrir el zip. El default es que si.

use_dict

Si descargar el diccionario de site.covid.dic.

quiet

(opcional) Variable para no mostrar mensajes

cache_datos

Direccion donde guardar los datos en memoria usando pins para no tener que volver a descargarlos si nada ha cambiado

use_cache_on_failure

Booleana. Establece que si no se pueden descargar datos nuevos utilice los que tenga en memoria.

cache_diccionario

Direccion donde guardar el diccionario en memoria usando pins para no tener que volver a descargarlo si nada ha cambiado

force_download

analiza si cambio el pin y descarga datos nuevos en caso afirmativo

show_warnings

si arrojar warnings

board_url_name

Establece el nombre del pins::board_url para los datos abiertos

board_url_name_dict

Establece el nombre del pins::board_url para los datos abiertos

download_file_args

Lista de argumentos adicionales para download.file de los datos si se elige este metodo para descargar.

descarga_db_diccionario_ssa_args

Lista con argumentos adicionales para el pins::pin_download de datos abiertos

prestatement

Opciones adicionales para MariaDB yo pongo las que me han funcionado en que sea mas veloz la carta.

additional_mysqlimport_flags

La instalacion se hace con el siguiente comando para MariaDB:

mysqlimport --default-character-set=UTF8 --fields-terminated-by=',' --ignore-lines=1
--fields-enclosed-by='\"' --lines-terminated-by='\n' --user={user}" --password={password}
--use-threads={nthreads} {silent_flag} {additional_mysqlimport_flags} {dbname} ./{tblname}.csv

load_data_infile_options

Lista de opciones por si la importacion con mysqlimport no funciona se use un LOAD DATA INFILE como aquí https://dev.mysql.com/doc/refman/8.0/en/load-data.html Las variables se pasan como lista asi por ejemplo si queremos agregar la variable local se pone list("LOCAL" = "LOCAL") y si no la queremos list("LOCAL" = "").

datos_abiertos_unzipped_path

Camino a los datos abiertos csv si ya los descargaste y descomprimiste el archivo zip en un csv

Valor

List of values:

dats - Database table (if MARIADB) or database in tibble (if tibble)

con - Database connection (if MARIADB) or NULL (if tibble)

dict - List of tibbles containing the whole dictionary

Ejemplos

if (FALSE) {
#Lee los datos de MariaDB una vez descargados
descarga_datos_abiertos()
datos_covid   <- read_datos_abiertos()

#Es lo mismo que:
datos_covid   <- read_datos_abiertos_MariaDB()

#Descarga los datos y lee de un zip
direccion_zip <- descarga_db_datos_abiertos_tbl()
datos_covid   <- read_datos_abiertos(direccion_zip)

#Es lo mismo que:
datos_covid   <- read_datos_abiertos_zip(direccion_zip)

#Descarga los datos y lee de un csv
direccion_zip <- descarga_db_datos_abiertos_tbl()
direccion_csv <- unzip_db_datos_abiertos_tbl(direccion_zip)
datos_covid   <- read_datos_abiertos(direccion_csv)

datos_covid   <- read_datos_abiertos_csv(direccion_csv)
}