RT: Número efectivo de reproducción

rt Calcula el número efectivo de reproducción por fecha y entidad

Uso

rt(
  datos_covid = NULL,
  entidades = c("AGUASCALIENTES", "BAJA CALIFORNIA", "BAJA CALIFORNIA SUR", "CAMPECHE",
    "CHIAPAS", "CHIHUAHUA", "CIUDAD DE MÉXICO", "COAHUILA DE ZARAGOZA", "COLIMA",
    "DURANGO", "GUANAJUATO", "GUERRERO", "HIDALGO", "JALISCO", "MÉXICO",
    "MICHOACÁN DE OCAMPO", "MORELOS", "NAYARIT", "NUEVO LEÓN", "OAXACA", "PUEBLA",
    "QUERÉTARO", "QUINTANA ROO", "SAN LUIS POTOSÍ", "SINALOA", "SONORA", "TABASCO",
    "TAMAULIPAS", "TLAXCALA", "VERACRUZ DE IGNACIO DE LA LLAVE", "YUCATÁN", "ZACATECAS"),
  group_by_entidad = TRUE,
  entidad_tipo = c("Unidad Medica", "Residencia", "Nacimiento"),
  fecha_tipo = c("Sintomas", "Ingreso", "Defuncion"),
  tipo_clasificacion = c("Sospechosos", "Confirmados COVID", "Negativo a COVID",
    "Inválido", "No realizado"),
  tipo_paciente = c("AMBULATORIO", "HOSPITALIZADO", "NO ESPECIFICADO"),
  list_name = "rt",
  min_date = as.Date("2021/11/21", format = "%Y/%m/%d"),
  method = "parametric_si",
  config = EpiEstim::make_config(list(mean_si = 3.5, std_si = 1.5)),
  ...
)

Argumentos

datos_covid

If no data is available it automatically downloads COVID-19 information.

entidades

Vector con las entidades de las unidades medicas a analizar. Opciones: AGUASCALIENTES, BAJA CALIFORNIA, BAJA CALIFORNIA SUR, CAMPECHE, CHIAPAS, CHIHUAHUA, CIUDAD DE MEXICO, COAHUILA DE ZARAGOZA , COLIMA, DURANGO, GUANAJUATO, GUERRERO, HIDALGO, JALISCO, MEXICO, MICHOACAN DE OCAMPO, MORELOS,NAYARIT NUEVO LEON, OAXACA ,PUEBLA, QUERETARO,QUINTANA ROO, SAN LUIS POTOSI, SINALOA, SONORA, TABASCO, TAMAULIPAS,TLAXCALA, VERACRUZ DE IGNACIO DE LA LLAVE, YUCATAN, ZACATECAS

group_by_entidad

Si junta las entidades en una sola o bien las muestra por separado sin agrupar.

entidad_tipo

Selecciona Unidad Medica, Nacimiento o Residencia. por default incluye Unidad Medica

fecha_tipo

Selecciona Ingreso, Sintomas o Defuncion por default incluye fecha de Sintomas

tipo_clasificacion

Vector con el tipo de clasificaciones a incluir: Sospechosos,Confirmados COVID, Negativo a COVID, Inv\u00e1lido, No realizado

tipo_paciente

Vector con el tipo de pacientes a incluir. Opciones: AMBULATORIO, HOSPITALIZADO, NO ESPECIFICADO

list_name

Asigna un nombre en la lista de datos a la base generada

min_date

Mínima fecha a partir de la cual estimar el RT

method

Method for estimating RT with EpiEstim::estimate_R

config

Configuration for estimating RT with EpiEstim::estimate_R

...

Parámetros adicionales para EpiEstim::estimate_R.

Valor

Appends a la lista de datos_covid una nueva entrada de nombre list_name

(default: casos) con una base de datos (tibble) con los resultados agregados.

positividad - Base de datos generara con los datos agregados (el nombre cambia si se usa list_name).

dict - Diccionario de datos

dats - Datos originales (conexion a DB)

disconnect - Funcion para desconectarte de DB

Detalles

Por default calcula el número efectivo de reproducción por estado This is not an official product / este no es un producto oficial

Ejemplos

if (FALSE) {
datos_covid <- descarga_datos_abiertos(language = "Espanol")

#Casos a nivel nacional en los confirmados
datos_covid <- datos_covid %>%
        rt(tipo_clasificacion == "Confirmados COVID",
          group_by_entidad = FALSE)

#Casos en AGS, CHI en los confirmados
datos_covid <- datos_covid %>%
        rt(entidades = c("CHIHUAHUA","AGUASCALIENTES"),
          tipo_clasificacion == "Confirmados COVID",
          group_by_entidad = TRUE)
}