CASOS

casos Calcula el numero de casos registrados por fecha agrupando (o sin hacerlo) por covariables.

Uso

casos(
  datos_covid,
  entidades = c("AGUASCALIENTES", "BAJA CALIFORNIA", "BAJA CALIFORNIA SUR", "CAMPECHE",
    "CHIAPAS", "CHIHUAHUA", "CIUDAD DE MÉXICO", "COAHUILA DE ZARAGOZA", "COLIMA",
    "DURANGO", "GUANAJUATO", "GUERRERO", "HIDALGO", "JALISCO", "MÉXICO",
    "MICHOACÁN DE OCAMPO", "MORELOS", "NAYARIT", "NUEVO LEÓN", "OAXACA", "PUEBLA",
    "QUERÉTARO", "QUINTANA ROO", "SAN LUIS POTOSÍ", "SINALOA", "SONORA", "TABASCO",
    "TAMAULIPAS", "TLAXCALA", "VERACRUZ DE IGNACIO DE LA LLAVE", "YUCATÁN", "ZACATECAS"),
  group_by_entidad = TRUE,
  entidad_tipo = c("Unidad Medica", "Residencia", "Nacimiento"),
  fecha_tipo = c("Sintomas", "Ingreso", "Defuncion"),
  tipo_clasificacion = c("Sospechosos", "Confirmados COVID", "Negativo a COVID",
    "Inválido", "No realizado"),
  group_by_tipo_clasificacion = FALSE,
  tipo_paciente = c("AMBULATORIO", "HOSPITALIZADO", "NO ESPECIFICADO"),
  group_by_tipo_paciente = FALSE,
  tipo_uci = c("SI", "NO", "NO APLICA", "SE IGNORA", "NO ESPECIFICADO"),
  group_by_tipo_uci = FALSE,
  tipo_sector = c("CRUZ ROJA", "DIF", "ESTATAL", "IMSS", "IMSS-BIENESTAR", "ISSSTE",
    "MUNICIPAL", "PEMEX", "PRIVADA", "SEDENA", "SEMAR", "SSA", "UNIVERSITARIO",
    "NO ESPECIFICADO"),
  group_by_tipo_sector = FALSE,
  defunciones = FALSE,
  edad_cut = NULL,
  as_tibble = TRUE,
  fill_zeros = as_tibble,
  list_name = "casos",
  .grouping_vars = c()
)

Argumentos

datos_covid

If no data is available it automatically downloads COVID-19 information.

entidades

Vector con las entidades de las unidades medicas a analizar. Opciones: AGUASCALIENTES, BAJA CALIFORNIA, BAJA CALIFORNIA SUR, CAMPECHE, CHIAPAS, CHIHUAHUA, CIUDAD DE MEXICO, COAHUILA DE ZARAGOZA , COLIMA, DURANGO, GUANAJUATO, GUERRERO, HIDALGO, JALISCO, MEXICO, MICHOACAN DE OCAMPO, MORELOS,NAYARIT NUEVO LEON, OAXACA ,PUEBLA, QUERETARO,QUINTANA ROO, SAN LUIS POTOSI, SINALOA, SONORA, TABASCO, TAMAULIPAS,TLAXCALA, VERACRUZ DE IGNACIO DE LA LLAVE, YUCATAN, ZACATECAS

group_by_entidad

Si junta las entidades en una sola o bien las muestra por separado sin agrupar.

entidad_tipo

Selecciona Unidad Medica, Nacimiento o Residencia. por default incluye Unidad Medica

fecha_tipo

Selecciona Ingreso, Sintomas o Defuncion por default incluye fecha de Sintomas

tipo_clasificacion

Vector con el tipo de clasificaciones a incluir: Sospechosos,Confirmados COVID, Negativo a COVID, Inv\u00e1lido, No realizado

group_by_tipo_clasificacion

Boolean determinando si regresa la base con cada entrada agrupada por tipo de clasificación (es decir cada fecha se generan tantos observaciones como grupos de tipo de clasificación)

tipo_paciente

Vector con el tipo de pacientes a incluir. Opciones: AMBULATORIO, HOSPITALIZADO, NO ESPECIFICADO

group_by_tipo_paciente

Boolean determinando si regresa la base con cada entrada agrupada por tipo de paciente (es decir cada fecha se generan tantos observaciones como grupos de tipo de paciente)

tipo_uci

Vector con el tipo de valores para Unidad de Cuidado Intensivo a incluir: SI,NO,NO APLICA,SE IGNORA,NO ESPECIFICADO

group_by_tipo_uci

Boolean determinando si regresa la base con cada entrada agrupada por tipo de uci (es decir cada fecha se generan tantos observaciones como grupos de tipo de uci)

tipo_sector

Vector con los sectores del sistema de salud a incluir: CRUZ ROJA,DIF,ESTATAL,IMSS,IMSS-BIENESTAR,ISSSTE, MUNICIPAL,PEMEX, PRIVADA,SEDENA,SEMAR,SSA, UNIVERSITARIO,NO ESPECIFICADO.

group_by_tipo_sector

Boolean determinando si regresa la base con cada entrada agrupada por tipo de sector (es decir cada fecha se generan tantos observaciones como grupos de tipo de sector)

defunciones

Boolean si incluir sólo defunciones TRUE o a todos FALSE.

edad_cut

Vector con secuencia de edades para hacer grupos. Por ejemplo edad_cut = c(0, 10, Inf) arma dos grupos de edad de 0 a 10 y de 10 a infinito o bien edad_cut = c(15, 20) deja sólo los registros entre 15 y 20 años. Por default es NULL

as_tibble

Regresar como tibble el resultado. En caso de que as_tibble sea FALSE se devuelve como conexion en MARIADB.

fill_zeros

En caso de que el resultado sea un tibble regresa observaciones para todas las combinaciones de variables incluyendo como 0 donde no se observaron casos. En caso contrario no se incluyen las filas donde no se observaron casos.

list_name

Asigna un nombre en la lista de datos a la base generada

.grouping_vars

Vector de variables adicionales de agrupacion de los conteos

Valor

Appends a la lista de datos_covid una nueva entrada de nombre list_name

(default: casos) con una base de datos (tibble o dbConnection) con los resultados agregados.

casos - Base de datos generara con los datos agregados (el nombre cambia si se usa list_name).

dict - Diccionario de datos

dats - Datos originales (conexion a DB)

disconnect - Función para desconectarte de DB

Detalles

Por default calcula el total de casos.

Ejemplos

if (FALSE) {
datos_covid <- descarga_datos_abiertos(language = "Espanol")

#Casos a nivel nacional
datos_covid <- datos_covid %>% casos()
head(datos_covid$casos)

#Defunciones nacional
defunciones <- datos_covid %>% casos(defunciones = TRUE, list_name = "Defunciones")

#Positivos en Jalisco y Colima
casos_col_jal <- datos_covid %>%
                    casos(entidades = c("JALISCO","COLIMA"), list_name = "Jaliscolima")

#Agrupando casos por tipo de clasificacion
confirmados <- datos_covid %>%
    casos(entidades = c("JALISCO","COLIMA"),
          group_by_tipo_clasificacion = TRUE, list_name = "Jaliscolima2")

#Regresa la suma de los de COLIMA + JALISCO
casos_col_jal <- datos_covid %>%
    casos(entidades = c("JALISCO","COLIMA"),
          group_by_tipo_clasificacion = TRUE,
          tipo_paciente = c("AMBULATORIO", "HOSPITALIZADO"),
          group_by_tipo_paciente = TRUE,
          list_name = "Jalisco + colima")

#Si deseas agrupar por una variable que no este en las opciones
casos_col_jal <- datos_covid %>%
    casos(entidades = c("JALISCO","COLIMA"),
          group_by_tipo_clasificacion = TRUE,
          tipo_paciente = c("AMBULATORIO", "HOSPITALIZADO"),
          group_by_tipo_paciente = TRUE,
          list_name = "Jalisco + colima diabetes",
          .grouping_vars = c("DIABETES"))
}