Descarga de datos abiertos

Conjunto de funciones para la descarga de datos abiertos de la Direccion General de Epidemiologia

La funcion de descarga principal es descarga_datos_abiertos llama las siguientes funciones en orden:

descarga_diccionario()

descarga_db()

pega_db_datos_abiertos_tbl_y_diccionario()

A su vez descarga_diccionario ejecuta las siguientes para obtener el diccionario de datos:

descarga_db_diccionario_ssa()

unzip_db_diccionario_ssa()

parse_db_diccionario_ssa()

A su vez descarga_db ejecuta las siguientes para obtener los datos abiertos:

descarga_db_datos_abiertos_tbl()

unzip_db_datos_abiertos_tbl()

parse_db_datos_abiertos_tbl()

Si en algun momento se interrumpio la descarga o hubo problemas de conexion o detuviste el proceso de generacion de la base de datos abiertos puedes llamar a las funciones de read_datos_abiertos()

Uso

descarga_datos_abiertos(
  read_format = c("MariaDB", "tibble"),
  user = Sys.getenv("MariaDB_user"),
  password = Sys.getenv("MariaDB_password"),
  dbname = Sys.getenv("MariaDB_dbname"),
  host = Sys.getenv("MariaDB_host"),
  group = Sys.getenv("MariaDB_group"),
  port = Sys.getenv("MariaDB_port"),
  tblname = "covidmx",
  nthreads = max(parallel::detectCores() - 1, 1),
  download_process = c("pins", "download.file"),
  site.covid = paste0("http://datosabiertos.salud.gob.mx",
    "/gobmx/salud/datos_abiertos/datos", "_abiertos_covid19.zip"),
  site.covid.dic = paste0("http://datosabiertos.salud.", "gob.mx/gobmx/salud/datos_a",
    "biertos/diccionario_datos_", "covid19.zip"),
  unzip_command = Sys.getenv("unzip_command"),
  unzip_args = Sys.getenv("unzip_args"),
  unzip_args_dict = list(exdir = ".", overwrite = TRUE),
  check_unzip_install = TRUE,
  clear_zip = download_process[1] != "pins",
  clear_csv = TRUE,
  use_dict = TRUE,
  datos_abiertos_zip_path = NULL,
  datos_abiertos_unzipped_path = NULL,
  datos_abiertos_tbl = NULL,
  diccionario_zip_path = NULL,
  diccionario_unzipped_path = NULL,
  diccionario = NULL,
  quiet = FALSE,
  cache_datos = NULL,
  use_cache_on_failure = TRUE,
  cache_diccionario = NULL,
  force_download = FALSE,
  show_warnings = TRUE,
  board_url_name = "datos_abiertos",
  board_url_name_dict = "diccionario_covid",
  download_file_args = list(method = "curl", destfile = tempfile(), quiet = quiet),
  descarga_db_datos_abiertos_tbl_args = list(),
  descarga_db_diccionario_ssa_args = list(),
  prestatement = paste0("SET sql_mode = 'NO_ENGINE_SUBSTITUTION,",
    "NO_AUTO_CREATE_USER';"),
  additional_mysqlimport_flags = c("--local"),
  load_data_infile_options = list(LOW_PRIORITY = "", CONCURRENT = "", LOCAL = "LOCAL",
    PARTITION = ""),
  ...
)

descarga_db(
  read_format = c("MariaDB", "tibble"),
  user = Sys.getenv("MariaDB_user"),
  password = Sys.getenv("MariaDB_password"),
  dbname = Sys.getenv("MariaDB_dbname"),
  host = Sys.getenv("MariaDB_host"),
  group = Sys.getenv("MariaDB_group"),
  port = Sys.getenv("MariaDB_port"),
  tblname = "covidmx",
  nthreads = max(parallel::detectCores() - 1, 1),
  download_process = c("pins", "download.file"),
  site.covid = paste0("http://datosabiertos.salud.gob.mx",
    "/gobmx/salud/datos_abiertos/datos", "_abiertos_covid19.zip"),
  unzip_command = Sys.getenv("unzip_command"),
  unzip_args = Sys.getenv("unzip_args"),
  check_unzip_install = TRUE,
  clear_zip = download_process[1] != "pins",
  clear_csv = TRUE,
  force_download = FALSE,
  show_warnings = TRUE,
  datos_abiertos_zip_path = NULL,
  datos_abiertos_unzipped_path = NULL,
  datos_abiertos_tbl = NULL,
  quiet = FALSE,
  board_url_name = "datos_abiertos",
  cache = NULL,
  use_cache_on_failure = TRUE,
  download_file_args = list(method = "curl", destfile = tempfile(), quiet = quiet),
  descarga_db_datos_abiertos_tbl_args = list(),
  prestatement = paste0("SET sql_mode = 'NO_ENGINE_SUBSTITUTION,",
    "NO_AUTO_CREATE_USER';"),
  additional_mysqlimport_flags = c("--local"),
  load_data_infile_options = list(LOW_PRIORITY = "", CONCURRENT = "", LOCAL = "LOCAL",
    PARTITION = ""),
  ...
)

descarga_diccionario(
  download_process = c("pins", "download.file"),
  site.covid.dic = paste0("http://datosabiertos.salud.", "gob.mx/gobmx/salud/datos_a",
    "biertos/diccionario_datos_", "covid19.zip"),
  quiet = FALSE,
  clear_zip = download_process[1] != "pins",
  clear_csv = TRUE,
  diccionario_zip_path = NULL,
  diccionario_unzipped_path = NULL,
  diccionario = NULL,
  board_url_name_dict = "diccionario_covid",
  cache_diccionario = NULL,
  use_cache_on_failure = TRUE,
  force_download = FALSE,
  show_warnings = TRUE,
  download_file_args_dict = list(method = "curl", destfile = tempfile(), quiet = quiet),
  unzip_args_dict = list(exdir = ".", overwrite = TRUE),
  descarga_db_diccionario_ssa_args = list()
)

descarga_db_datos_abiertos_tbl(
  download_process = c("pins", "download.file"),
  site.covid = paste0("http://datosabiertos.salud.gob.mx",
    "/gobmx/salud/datos_abiertos/datos", "_abiertos_covid19.zip"),
  quiet = FALSE,
  board_url_name = "datos_abiertos",
  cache = NULL,
  use_cache_on_failure = TRUE,
  force_download = FALSE,
  show_warnings = TRUE,
  download_file_args = list(method = "curl", destfile = tempfile(), quiet = quiet),
  ...
)

descarga_db_diccionario_ssa(
  download_process = c("pins", "download.file"),
  site.covid.dic = paste0("http://datosabiertos.salud.", "gob.mx/gobmx/salud/datos_a",
    "biertos/diccionario_datos_", "covid19.zip"),
  quiet = FALSE,
  board_url_name_dict = "diccionario_covid",
  cache_diccionario = NULL,
  use_cache_on_failure = TRUE,
  force_download = FALSE,
  show_warnings = TRUE,
  download_file_args_dict = list(method = "curl", destfile = tempfile(), quiet = quiet),
  ...
)

unzip_db_datos_abiertos_tbl(
  datos_abiertos_zip_path,
  unzip_command = Sys.getenv("unzip_command"),
  unzip_args = Sys.getenv("unzip_args"),
  check_unzip_install = TRUE,
  quiet = FALSE,
  clear_zip = FALSE
)

unzip_db_diccionario_ssa(
  diccionario_zip_path,
  unzip_args_dict = list(exdir = ".", overwrite = TRUE),
  clear_zip = FALSE
)

parse_db_diccionario_ssa(diccionario_unzipped_path, clear_csv = FALSE)

parse_db_datos_abiertos_tbl(
  datos_abiertos_unzipped_path,
  read_format = c("MariaDB", "tibble"),
  user = Sys.getenv("MariaDB_user"),
  password = Sys.getenv("MariaDB_password"),
  dbname = Sys.getenv("MariaDB_dbname"),
  host = Sys.getenv("MariaDB_host"),
  group = Sys.getenv("MariaDB_group"),
  port = Sys.getenv("MariaDB_port"),
  tblname = "covidmx",
  nthreads = max(parallel::detectCores() - 1, 1),
  quiet = TRUE,
  clear_csv = FALSE,
  prestatement = paste0("SET sql_mode = 'NO_ENGINE_SUBSTITUTION,",
    "NO_AUTO_CREATE_USER';"),
  additional_mysqlimport_flags = c("--local"),
  load_data_infile_options = list(LOW_PRIORITY = "", CONCURRENT = "", LOCAL = "LOCAL",
    PARTITION = ""),
  ...
)

pega_db_datos_abiertos_tbl_y_diccionario(datos_abiertos_tbl, diccionario)

Argumentos

read_format

"MariaDB" o "tibble" establece el formato de lectura de la base de datos. En la mayoria de los casos "tibble" va a resultar en un error de memoria.

user

Usuario para dbConnect i.e. tu usuario de MariaDB. Puedes ponerlo como variable ambiental con Sys.setenv("MariaDB_user" = "tu_usuario")

password

password para dbConnect i.e. tu password de MariaDB Puedes ponerlo como variable ambiental con Sys.setenv("MariaDB_password" = "tu_passwodrd")

dbname

Nombre de la base de datos DATABASE paradbConnect puedes ponerlo con Sys.setenv("MariaDB_dbname" = "covidmx")

host

(opcional) Host para dbConnect. Si no sabes cual tienes seguro es localhost. Puedes ponerlo con Sys.setenv("MariaDB_host" = "localhost")

group

(opcional) Grupo de dbConnect i.e. tu grupo de MariaDB. Ponlo como variable ambiental usando Sys.setenv("MariaDB_group" = "tu_grupo")

port

Puerto de la conexion con MariaDB. Puedes ponerlo con Sys.setenv("MariaDB_group" = "tu_grupo")

tblname

Nombre de la tabla dentro de el DATABASE dbname donde guardar los datos por default se llama covidmx.

nthreads

Número de threads para escribir el archivo en MariaDB.

download_process

(opcional) Metodo para descargar ya sea pins o download.file. Se recomienda pins pues guarda en memoria la fecha de la ultima descarga y analiza si ha pasado mas de un dia desde la descarga. En caso afirmativo verifica si el archivo ha cambiado y si hubo cambios entonces lo descarga.

site.covid

Sitio web con el vinculo al archivo .zip de los datos abiertos. Puedes cambiarlo por uno de los historicos, por ejemplo.

site.covid.dic

Sitio para descarga del diccionario de datos.

unzip_command

Forma de extraer la base de datos de datos abiertos. La forma de llamarla es con system2(unzip_command, args = c(unzip_args, file_download_data)).

unzip_args

Argumentos de extraccion de la base de datos de datos abiertos. La forma de llamarla es con system2(unzip_command, args = c(unzip_args, file_download_data)).

unzip_args_dict

Lista de argumentos para usar utils::unzip en el diccionario de datos.

check_unzip_install

Bandera de verificacion para checar si tienes lo necesario para unzippear los datos.

clear_zip

Si borrar los archivos .zip descargados para el diccionario y los datos abiertos. No se recomienda si estas usando pins. Ve la nota para mas informacion.

clear_csv

Si borrar los archivos .csv que se generan despues de abrir el zip. El default es que si.

use_dict

Si descargar el diccionario de site.covid.dic.

datos_abiertos_zip_path

Camino a los datos abiertos si ya los descargaste en zip

datos_abiertos_unzipped_path

Camino a los datos abiertos csv si ya los descargaste y descomprimiste el archivo zip en un csv

datos_abiertos_tbl

Lo que resulta de realizar una descarga de los datos abiertos usando descarga_db

diccionario_zip_path

Camino al diccionario si ya losdescargaste en zip

diccionario_unzipped_path

Camino al diccionario csv si ya lo descargaste y descomprimiste el archivo zip en un csv

diccionario

Lo que resulta de realizar una descarga del diccionario usando descarga_diccionario

quiet

(opcional) Variable para no mostrar mensajes

cache_datos

Direccion donde guardar los datos en memoria usando pins para no tener que volver a descargarlos si nada ha cambiado

use_cache_on_failure

Booleana. Establece que si no se pueden descargar datos nuevos utilice los que tenga en memoria.

cache_diccionario

Direccion donde guardar el diccionario en memoria usando pins para no tener que volver a descargarlo si nada ha cambiado

force_download

analiza si cambio el pin y descarga datos nuevos en caso afirmativo

show_warnings

si arrojar warnings

board_url_name

Establece el nombre del pins::board_url para los datos abiertos

board_url_name_dict

Establece el nombre del pins::board_url para los datos abiertos

download_file_args

Lista de argumentos adicionales para download.file de los datos si se elige este metodo para descargar.

descarga_db_datos_abiertos_tbl_args

Lista con argumentos adicionales para el pins::pin_download de datos abiertos

descarga_db_diccionario_ssa_args

Lista con argumentos adicionales para el pins::pin_download de datos abiertos

prestatement

Opciones adicionales para MariaDB yo pongo las que me han funcionado en que sea mas veloz la carta.

additional_mysqlimport_flags

La instalacion se hace con el siguiente comando para MariaDB:

mysqlimport --default-character-set=UTF8 --fields-terminated-by=',' --ignore-lines=1
--fields-enclosed-by='\"' --lines-terminated-by='\n' --user={user}" --password={password}
--use-threads={nthreads} {silent_flag} {additional_mysqlimport_flags} {dbname} ./{tblname}.csv

load_data_infile_options

Lista de opciones por si la importacion con mysqlimport no funciona se use un LOAD DATA INFILE como aquí https://dev.mysql.com/doc/refman/8.0/en/load-data.html Las variables se pasan como lista asi por ejemplo si queremos agregar la variable local se pone list("LOCAL" = "LOCAL") y si no la queremos list("LOCAL" = "").

...

Parametros adicionales para pins::pin_download

cache

parametro para el cache de pins::board_url

download_file_args_dict

Lista de argumentos adicionales para download.file del diccionario si se elige este metodo de descarga.

Valor

List of values:

dats - Tabla conectada mediante DBI::dbConnect (si MariaDB) o tibble (si tibble)

disconnect - Funcion para cerrar la conexion a la base de datos.

dict - Lista de tibbles con el diccionario de datos para cada variable

Detalles

Necesitas tener una instalacion funcionando de MariaDB. El programa se encarga de descargar la base de datos de la DGE, abrir el archivo .zip y crear una tabla de nombre covidmx dentro de tu database dbname. Asegurate de que tu usuario user tenga permisos de escritura. Para mas informacion sobre instalacion y uso de MariaDB consulta el articulo correspondiente.

Si tienes RAM que te sobre puedes no crear una base de datos en MariaDB sino leer directo el archivo csv. Esto se logra con read_format = tibble. No lo recomiendo pues puedes terminar con tu sesion de R si se te acaba la memoria.

Windows Para abrir el archivo .zip requieres tambien descargar e instalar 7Zip por default el sistema lo busca en C:\\Program Files\\7-Zip\\7z.exe pero si no esta ese directorio es necesario que en unzip_command especifiques el camino donde se instalo 7z.exe.

Nota

No te recomiendo borrar el cache con clear_zip o editarlo por cualquier otro medio si estas usando pins pues puede romperse la dependencia. Si accidentalmente lo borraste usa pins::board_cache_path() para ir al path y borrar manualmente toda la carpeta.

Uso de pins

Para almacenar los datos se utiliza un pequenio cambio sobre la libreria pins. Los datos se descargan y se almacenan en cache junto con informacion sobre cuando fue la descarga. Si no ha pasado un dia desde la ultima descarga no se descarga nada nuevo. Si los datos que se tienen no han cambiado respecto a lo que esta en linea tampoco se vuelven a descargar aunque haya pasado mas de un dia.

Si se te fue el Internet No te preocupes, pins lee tu descarga mas reciente.

Para ver donde estan descargados tus datos usa pins::board_cache_path(). Para borrarlos usa pins::cache_prune().

Ver también

Ejemplos

if (FALSE) {
#Descarga de la base de datos junto con diccionario en MariaDB
datos_covid <- descarga_datos_abiertos()

#Luego haces algo con esos datos...

#Cuando terminas cierras la sesión:
datos_covid$disconnect()

#Descarga solo el diccionario
diccionario <- descarga_diccionario()

#O bien descarga solo los datos abiertos
datos_abiertos <- descarga_db()

#Pegalos en el formato que se necesita para el resto de funciones
datos_covid <- pega_db_datos_abiertos(datos_abiertos, diccionario)

#Tambien puedes descargar paso por paso
datos_abiertos <- descarga_db_datos_abiertos_tbl() %>% #Descarga
                     unzip_db_datos_abiertos_tbl() %>% #Unzippea
                     parse_db_datos_abiertos_tbl()     #MariaDB

#O bien el diccionario
diccionario <- descarga_db_diccionario_ssa()   %>% #Descarga
                    unzip_db_diccionario_ssa() %>% #Unzippea
                    parse_db_diccionario_ssa()     #Tibble
}