DESCARGA BASE DE DATOS ABIERTOS DE LA DIRECCION GENERAL DE EPIDEMIOLOGIA

descarga_datos_abiertos downloads and labels all COVID-19 data for Mexico from the Direccion General de EpidemiologIa at https://www.gob.mx/salud/documentos/datos-abiertos-152127.

descarga_datos_abiertos descarga y etiqueta todos los datos de COVID-19 de Mexico de la Direccion General de EpidemiologIa en https://www.gob.mx/salud/documentos/datos-abiertos-152127.

Usage

descarga_datos_abiertos(
  download_method = "curl",
  file_download_data = tempfile(),
  file_download_dictionary = tempfile(),
  remove_zip_after_download = TRUE,
  quiet = FALSE,
  parse_warnings = FALSE,
  language = c("English", "Español"),
  read_format = c("MariaDB", "tibble"),
  user = Sys.getenv("MariaDB_user"),
  password = Sys.getenv("MariaDB_password"),
  dbname = Sys.getenv("MariaDB_dbname"),
  host = Sys.getenv("MariaDB_host"),
  group = Sys.getenv("MariaDB_group"),
  port = Sys.getenv("MariaDB_port"),
  tblname = "covidmx",
  dict_file = "diccionario_covid.RData",
  download_dict = TRUE,
  save_dict = download_dict,
  site.covid = paste0("http://datosabiertos.salud.gob.mx/gobmx/salud",
    "/datos_abiertos/datos_abiertos_covid19.zip"),
  site.covid.dic = paste0("http://datosabiertos.salud.gob.mx/gobmx/salud/",
    "datos_abiertos/diccionario_datos_covid19.zip"),
  nthreads = max(parallel::detectCores() - 1, 1)
)

Arguments

download_method

Methods for download file (default = "curl"). Other options are "internal", "wininet" (Windows) "libcurl", "wget", "curl". See download.file / Metodos para descargar el archivo (default = "curl"). Otras opciones incluyen "interno","wininet" (Windows) " libcurl", "wget", "curl". Vease download.file.

file_download_data

Name of file to save the data / Nombre del archivo para guardar los datos.

file_download_dictionary

Name of file to save the dictionary / Nombre del archivo donde guardar el diccioanrio.

remove_zip_after_download

If the downloaded zip file should be saved / Si los archivos zip descargados deben ser almacenados.

quiet

Show messages? / Mostrar mensajes?

parse_warnings

If show parsing errors when reading csv / Si muestra errores cuando lee el csv

language

('Español', 'English') Message languages/ Idiomas del mensaje.

read_format

"MariaDB" or "tibble" establishes how the database should be stored. In most machines "tibble" will result in a memory error

user

User for dbConnect i.e. your MariaDB user

password

password for dbConnect i.e. your MariaDB password

dbname

Database name for dbConnect i.e. what database is going to be stored

host

Host for dbConnect i.e. your MariaDB host (usually localhost)

group

Group for dbConnect i.e. your MariaDB group (can be NULL)

port

Port connection for MariaDB

tblname

Name of table to save in MariaDB

dict_file

RData File where to save/read the dictionary.

download_dict

If download the dictionary from site.covid.dic. If download_dict is FALSE then dict_file must be specified to read the dictionary.

save_dict

If the downloaded dictionary is to be saved in file dict_file

site.covid

Site for download the covid information

site.covid.dic

Site for download the covid dictionary for the data

nthreads

Number of threads for writing to MariaDB.

Value

List of values:

dats - Database table (if MARIADB) or database in tibble (if tibble)

con - Database connection (if MARIADB) or NULL (if tibble)

dict - List of tibbles containing the whole dictionary

Details

This is not an official product / este no es un producto oficial

English While reading the data it is perfectly normal to get a Warning for parsing: the dataset has issues.

Spanish En la lectura de datos es normal tener un Warning por parsing: la base tiene errores de registro.

Necesitas tener una instalación funcionando de MariaDB. El programa se encarga de descargar la base de datos de la DGE, abrir el archivo .zip y crear una tabla de nombre covidmx dentro de tu database dbname. Asegúrate de que tu usuario user tenga permisos de escritura. Para más información sobre instalación y uso de MariaDB consulta la viñeta rellenar.

Si tienes RAM que te sobre puedes no crear una base de datos en MariaDB sino leer directo el archivo csv. Esto se logra con read_format = tibble. No lo recomiendo pues puedes terminar con tu sesión de R si se te acaba la memoria.

Examples

if (FALSE) {
datos_covid <- descarga_datos_abiertos(language = "Espanol")
}