POSITIVIDAD

positividad Calcula la positividad definida como el radio de $$\frac{Pruebas positivas}{Total de pruebas}$$

Usage

positividad(
  datos_covid = NULL,
  entidades = c("AGUASCALIENTES", "BAJA CALIFORNIA", "BAJA CALIFORNIA SUR", "CAMPECHE",
    "CHIAPAS", "CHIHUAHUA", "CIUDAD DE MÉXICO", "COAHUILA DE ZARAGOZA", "COLIMA",
    "DURANGO", "GUANAJUATO", "GUERRERO", "HIDALGO", "JALISCO", "MÉXICO",
    "MICHOACÁN DE OCAMPO", "MORELOS", "NAYARIT", "NUEVO LEÓN", "OAXACA", "PUEBLA",
    "QUERÉTARO", "QUINTANA ROO", "SAN LUIS POTOSÍ", "SINALOA", "SONORA", "TABASCO",
    "TAMAULIPAS", "TLAXCALA", "VERACRUZ DE IGNACIO DE LA LLAVE", "YUCATÁN", "ZACATECAS"),
  group_by_entidad = TRUE,
  entidad_tipo = c("Unidad Médica", "Residencia", "Nacimiento"),
  fecha_tipo = c("Síntomas", "Ingreso", "Defunción")
)

Arguments

datos_covid

If no data is available it automatically downloads COVID-19 information.

entidades

Vector con las entidades de las unidades medicas a analizar. Opciones: AGUASCALIENTES, BAJA CALIFORNIA, BAJA CALIFORNIA SUR, CAMPECHE, CHIAPAS, CHIHUAHUA, CIUDAD DE MEXICO, COAHUILA DE ZARAGOZA , COLIMA, DURANGO, GUANAJUATO, GUERRERO, HIDALGO, JALISCO, MEXICO, MICHOACAN DE OCAMPO, MORELOS,NAYARIT NUEVO LEON, OAXACA ,PUEBLA, QUERETARO,QUINTANA ROO, SAN LUIS POTOSI, SINALOA, SONORA, TABASCO, TAMAULIPAS,TLAXCALA, VERACRUZ DE IGNACIO DE LA LLAVE, YUCATAN, ZACATECAS

group_by_entidad

Si junta las entidades en una sola o bien las muestra por separado sin agrupar.

entidad_tipo

Selecciona Unidad Medica, Nacimiento o Residencia. por default incluye Unidad Medica

fecha_tipo

Selecciona Ingreso, Sintomas o Defuncion por default incluye fecha de Sintomas

Details

Por default calcula la positividad total en pruebas tanto de laboratorio como de antigeno vease positividad_antigeno y positividad_pcr para las positividades especificas de estas pruebas.

This is not an official product / este no es un producto oficial

Examples

if (FALSE) {
datos_covid <- descarga_datos_abiertos(language = "Espanol")

#Positividad nacional
datos_covid %>% positividad()

#Positividad en Jalisco y Colima
datos_covid %>% positividad(entidades = c("JALISCO","COLIMA")) %>%
         covid_plot()
datos_covid %>% positividad(entidades = c("JALISCO","COLIMA"),
                            group_by_entidad = T) %>%
         covid_plot()
}