LEE LA BASE DE DATOS ABIERTOS DE LA DIRECCION GENERAL DE EPIDEMIOLOGIA QUE YA DESCARGASTE

read_datos_abiertos Lee los datos abiertos almacenados en tu base de MariaDB que bajaste con descarga_datos_abiertos

Uso

read_datos_abiertos(
  user = Sys.getenv("MariaDB_user"),
  password = Sys.getenv("MariaDB_password"),
  dbname = Sys.getenv("MariaDB_dbname"),
  host = Sys.getenv("MariaDB_host"),
  group = Sys.getenv("MariaDB_group"),
  port = Sys.getenv("MariaDB_port"),
  tblname = "covidmx",
  dict_file = "diccionario_covid.RData",
  download_dict = FALSE,
  download_method = "curl",
  save_dict = download_dict,
  file_download_dictionary = tempfile(),
  remove_zip_after_download = TRUE,
  site.covid.dic = paste0("http://datosabiertos.salud.gob.mx/gobmx/salud/",
    "datos_abiertos/diccionario_datos_covid19.zip")
)

Argumentos

user

User for dbConnect i.e. your MariaDB user

password

password for dbConnect i.e. your MariaDB password

dbname

Database name for dbConnect i.e. what database is going to be stored

host

Host for dbConnect i.e. your MariaDB host (usually localhost)

group

Group for dbConnect i.e. your MariaDB group (can be NULL)

port

Port connection for MariaDB

tblname

Name of table to save in MariaDB

dict_file

RData File where to save/read the dictionary.

download_dict

If download the dictionary from site.covid.dic. If download_dict is FALSE then dict_file must be specified to read the dictionary.

download_method

Methods for download file (default = "curl"). Other options are "internal", "wininet" (Windows) "libcurl", "wget", "curl". See download.file / Metodos para descargar el archivo (default = "curl"). Otras opciones incluyen "interno","wininet" (Windows) " libcurl", "wget", "curl". Vease download.file.

save_dict

If the downloaded dictionary is to be saved in file dict_file

file_download_dictionary

Name of file to save the dictionary / Nombre del archivo donde guardar el diccioanrio.

remove_zip_after_download

If the downloaded zip file should be saved / Si los archivos zip descargados deben ser almacenados.

site.covid.dic

Site for download the covid dictionary for the data

Valor

List of values:

dats - Database table (if MARIADB) or database in tibble (if tibble)

con - Database connection (if MARIADB) or NULL (if tibble)

dict - List of tibbles containing the whole dictionary

Detalles

This is not an official product / este no es un producto oficial

Spanish Necesitas tener una instalación funcionando de MariaDB. El programa se encarga de leer la base de datos de la DGE que bajamos con descarga_datos_abiertos. Basta con leer la tabla de nombre covidmx dentro de tu database dbname. Asegúrate de que tu usuario user tenga permisos de lectura. Para más información sobre instalación y uso de MariaDB consulta la viñeta rellenar.

Ejemplos

if (FALSE) {
datos_covid <- read_datos_abiertos()
}